sat: import some ric and game data
