+++ config/saturn/game_user_syms.txt
@@ -113,8 +113,19 @@ _g_ButtonMask = 0x060854C8;
 _g_LevelHPIncrease = 0x060854E0;
 _D_801375C8 = 0x06085524;
 _g_EquipMenuHelper = 0x0608554C;
-_DAT_060855ac = 0x060855AC;
-_DAT_06085c70 = 0x06085C70;
+_g_MenuPageScrollActive = 0x060855AC;
+_D_800A37F4 = 0x06085C30;
+_D_800A37FC = 0x06085C38;
+_D_06085C40 = 0x06085C40;
+_D_800A380C = 0x06085C48;
+_D_800A3814 = 0x06085C50;
+_D_800A381C = 0x06085C58;
+_D_800A3824 = 0x06085C60;
+_D_800A382C = 0x06085C68;
+_g_JosephsCloakColors = 0x06085C70;
+_D_800A3804 = 0x06085C78;
+_D_800A3834 = 0x06085C80;
+_g_EquippedCapePalette = 0x06085C88;
 _g_LuckCode = 0x06085C90;
 _g_AxeArmorCode = 0x06085C94;
 _g_GTIClubCode = 0x06085C98;
