+++ config/saturn/richter_user_syms.txt
@@ -8,6 +8,8 @@ _GetTeleportToOtherCastle = 0x060A5060;
 _func_80156DE4 = 0x060A5154;
 _D_80154568 = 0x060BED70;
 _RicInit = 0x060A5208;
+_D_060C2FA4 = 0x060C2FA4;
+_D_060C2FE8 = 0x060C2FE8;
 _CheckStageCollision = 0x060A5518;
 _CheckBladeDashInput = 0x060A5864;
 _CheckHighJumpInput = 0x060A5988;
