+++ src/saturn/game/btnmask.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s16 g_ButtonMask[] = {
+    0x0400, 0x0100, 0x0200, 0x0040, 0x0020, 0x0010, 0x0080, 0x0008,
+};
