+++ src/saturn/game/capecon.c
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u16 D_800A3804[4] = {0xA4A8, 0xAD0A, 0xCC21, 0xFC42};
+u16 D_800A3834[4] = {0xCC21, 0xFC42, 0xA4A8, 0xAD0A};
