+++ src/saturn/game/capepal.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u16 g_EquippedCapePalette[4] = {0x994A, 0xA5CE, 0x8421, 0x9CE7};
