+++ src/saturn/game/capetbl.c
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u16 D_800A37F4[4] = {0x994A, 0xA5CE, 0x8421, 0x9CE7};
+u16 D_800A37FC[4] = {0x8420, 0x8427, 0x8421, 0x9CE7};
+u16 D_06085C40[4] = {0x8000, 0x8000, 0x8000, 0x8000};
+u16 D_800A380C[4] = {0x910C, 0x996E, 0x8048, 0x80B2};
+u16 D_800A3814[4] = {0x90A1, 0x9102, 0xA821, 0xCC42};
+u16 D_800A381C[4] = {0x90C9, 0x94EB, 0x85E4, 0xA2CD};
+u16 D_800A3824[4] = {0x842C, 0x8C6F, 0x8018, 0x98DD};
+u16 D_800A382C[4] = {0x994A, 0xA5CE, 0x8000, 0xBC0F};
