+++ src/saturn/game/cfgjp.c
@@ -1,0 +1,949 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+// clang-format off
+
+SubweaponDef g_SubwpnDefs[] = {
+    {0, 30000, 0x0000, 0, 0x00, 0, 0x55, 0x00, 2, 0x01, 0, 0, 0},
+    {5, 1, 0x1040, 3, 0x04, 4, 0x55, 0x03, 2, 0x84, 0, 0, 0},
+    {10, 4, 0x1040, 2, 0x20, 4, 0x55, 0x09, 2, 0x83, 0, 0, 0},
+    {1, 3, 0x0020, 3, 0x20, 8, 0xB4, 0x1B, 2, 0x01, 11, 0, 0},
+    {5, 100, 0x1000, 1, 0x20, 4, 0xC8, 0x06, 2, 0x01, 12, 0, 0},
+    {8, 5, 0x1000, 1, 0x18, 4, 0x55, 0x4D, 2, 0x01, 0, 0, 0},
+    {0, 20, 0x1000, 1, 0x32, 4, 0xB2, 0x47, 2, 0x01, 0, 0, 0},
+    {8, 2, 0x1020, 3, 0x20, 4, 0x55, 0x18, 2, 0x01, 0, 0, 0},
+    {8, 3, 0x1000, 2, 0x0A, 8, 0xB4, 0x1A, 2, 0x01, 0, 0, 0},
+    {10, 5, 0x5000, 1, 0x0A, 4, 0x55, 0x16, 2, 0x01, 0, 0, 0},
+    {15, 10, 0x1000, 1, 0x0A, 4, 0x55, 0x16, 2, 0x01, 0, 0, 0},
+    {8, 1, 0x9000, 3, 0x10, 8, 0x55, 0x1B, 2, 0x01, 0, 0, 0},
+    {40, 5, 0x1020, 1, 0x18, 8, 0x55, 0x06, 2, 0x01, 0, 0, 0},
+};
+
+Equipment g_EquipDefs[] = {
+    {_S("すで"), "装備なし（素手）", 0, 0, 0x0000, 3, 255, 0x0000, 27, 42, 0, 5, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 1, 1, 1, 0},
+    {_S("カプセルモンスター1"), "半魚人を召喚して攻撃する【消費】", 40, 0, 0x0000, 8, 40, 0x0082, 88, 50, 0, 0, 2048, 0, 1, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 153, 151, 1, 0},
+    {_S("カプセルモンスター2"), "コウモリを召喚して攻撃する【消費】", 25, 0, 0x0000, 8, 41, 0x0082, 88, 50, 0, 0, 2048, 0, 1, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 153, 147, 1, 0},
+    {_S("カプセルモンスター3"), "スケルトンを召喚して攻撃する【消費】", 30, 0, 0x0000, 8, 42, 0x0082, 88, 50, 0, 0, 2048, 0, 1, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 153, 150, 1, 0},
+    {_S("シールドロッド"), "盾と相性の良い魔道具", 11, 2, 0x0000, 4, 7, 0x0002, 64, 48, 0, 13, 32948, 0, 0, 20, 0, 0x00000000, 0x00007FFE, 0, 4, 2, 1, 5, 5, 1, 0},
+    {_S("レザーシールド"), "なめし皮の盾　防御力は低い", 1, 1, 0x0000, 9, 8, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000002, 0x00000000, 0, 0, 4, 1, 6, 6, 1, 0},
+    {_S("ナイトシールド"), "騎士の持つ盾", 1, 2, 0x0000, 9, 9, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000004, 0x00000000, 0, 0, 4, 1, 7, 7, 1, 0},
+    {_S("アイアンシールド"), "鋼鉄製の硬い盾", 1, 2, 0x0000, 9, 10, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000008, 0x00000000, 0, 0, 4, 1, 8, 8, 1, 0},
+    {_S("アックスアーマーの盾"), "アックスアーマーが愛用する盾", 1, 3, 0x0000, 9, 11, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000010, 0x00000000, 0, 0, 4, 1, 9, 9, 1, 0},
+    {_S("もんしょうの盾"), "ハートの紋章が描かれた盾", 1, 4, 0x0000, 9, 23, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000020, 0x00000000, 0, 0, 4, 1, 10, 10, 1, 0},
+    {_S("ダークシールド"), "暗黒の盾　触れるだけでダメージ", 10, 3, 0x0000, 9, 24, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 30, 0, 0x00000040, 0x00000000, 0, 4, 6, 1, 11, 11, 1, 0},
+    {_S("めがみの盾"), "女神の紋様の盾", 1, 4, 0x0000, 9, 25, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000080, 0x00000000, 0, 0, 4, 1, 12, 12, 1, 0},
+    {_S("シャーマンの盾"), "魔術の儀式に使われた盾", 1, 4, 0x0000, 9, 26, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000100, 0x00000000, 0, 0, 4, 1, 13, 13, 1, 0},
+    {_S("メディウサの盾"), "切り落とされたメディウサの首", 13, 1, 0x0200, 9, 27, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 30, 0, 0x00000200, 0x00000000, 0, 20, 6, 1, 14, 14, 1, 0},
+    {_S("ドクロの盾"), "戦士の紋章をかかげた盾", 1, 4, 0x0000, 9, 28, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000400, 0x00000000, 0, 0, 4, 1, 15, 15, 1, 0},
+    {_S("フレイムシールド"), "火炎属性の盾", 1, 4, 0x8000, 9, 29, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00000800, 0x00000000, 0, 0, 4, 1, 16, 16, 1, 0},
+    {_S("アルカードシールド"), "以前から愛用している盾。全ての攻撃に強い", 20, 5, 0x0000, 9, 52, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00001000, 0x00000000, 0, 0, 4, 1, 4, 4, 1, 0},
+    {_S("あかつきの剣"), "不死の軍隊を召喚できる魔剣【両手持ち】", 27, 0, 0x0040, 5, 30, 0x0006, 96, 48, 0, 13, 32768, 184, 0, 16, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 17, 17, 1, 0},
+    {_S("バーゼラルド"), "一般的な短剣　", 1, 0, 0x0040, 0, 31, 0x0007, 105, 48, 0, 5, 32949, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 18, 18, 1, 0},
+    {_S("ショートソード"), "短めの剣　", 2, 0, 0x0040, 0, 37, 0x000F, 161, 48, 0, 8, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 19, 19, 1, 0},
+    {_S("コンバットナイフ"), "隠密性の高いフラット処理した短剣", 24, 0, 0x0040, 0, 32, 0x0009, 119, 48, 0, 5, 32950, 182, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 20, 20, 32, 0},
+    {_S("トンファー"), "２回攻撃する打撃武器【両手持ち】", 10, 0, 0x0000, 5, 33, 0x000B, 133, 48, 0, 16, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 21, 21, 32, 0},
+    {_S("ウェアバスター"), "獣人によく効く", 18, 0, 0x0040, 0, 39, 0x0010, 168, 48, 0, 13, 32768, 197, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 22, 22, 32, 0},
+    {_S("レイピア"), "刃の長い、刺すための剣", 7, 0, 0x0040, 0, 38, 0x0010, 168, 48, 0, 13, 32768, 198, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 23, 23, 32, 0},
+    {_S("ボンバーコイン"), "２種の効果を持つコイン【消費】", 50, 0, 0x0000, 7, 20, 0x0082, 88, 50, 0, 0, 256, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 24, 24, 1, 0},
+    {_S("マジックミサイル"), "魔法の矢で攻撃【消費】", 30, 0, 0x0040, 8, 22, 0x0082, 88, 50, 0, 0, 1024, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 25, 25, 1, 0},
+    {_S("アカサビ"), "赤く錆びた刀、抜けないときがある【両手持ち】", -2, 0, 0x0140, 5, 35, 0x000D, 147, 48, 0, 13, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 26, 26, 32, 0},
+    {_S("タケミツ"), "竹で作られた模造刀【両手持ち】", 1, 0, 0x0000, 5, 36, 0x000E, 154, 48, 0, 13, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 27, 27, 32, 0},
+    {_S("ショテル"), "エチオピアの両刀の曲剣", 20, 0, 0x0040, 0, 34, 0x000C, 140, 48, 0, 13, 32768, 183, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 28, 28, 1, 0},
+    {_S("みかん"), "果汁たっぷりジューシーなミカン【消費】", 6, 0, 0x0000, 6, 255, 0x0084, 88, 128, 0, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 29, 29, 1, 0},
+    {_S("りんご"), "蜜の多いリンゴ【消費】", 9, 0, 0x0000, 6, 255, 0x0084, 88, 128, 1, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 30, 30, 1, 0},
+    {_S("バナナ"), "腐りかけが一番うまい【消費】", 7, 0, 0x0000, 6, 255, 0x0084, 88, 128, 2, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 31, 31, 1, 0},
+    {_S("ぶどう"), "大きな粒のマスカット【消費】", 5, 0, 0x0000, 6, 255, 0x0084, 88, 128, 3, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 32, 32, 1, 0},
+    {_S("いちご"), "おいしい苺【消費】", 10, 0, 0x0000, 6, 255, 0x0084, 88, 128, 4, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 33, 33, 1, 0},
+    {_S("パイナップル"), "南国で採れたジューシーな逸品【消費】", 8, 0, 0x0000, 6, 255, 0x0084, 88, 128, 5, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 34, 34, 1, 0},
+    {_S("ピーナッツ"), "食べるのにコツがいる【消費】", 50, 0, 0x0000, 6, 255, 0x0084, 88, 128, 6, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 35, 35, 1, 0},
+    {_S("ベニテングダケ"), "毒を持つキノコ【消費】", 0, 0, 0x0000, 6, 255, 0x0084, 88, 128, 7, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 36, 36, 1, 0},
+    {_S("うまいキノコ"), "独特の芳香が食欲をそそるキノコ【消費】", 30, 0, 0x0000, 6, 255, 0x0084, 88, 128, 8, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 37, 37, 1, 0},
+    {_S("チーズケーキ"), "上品な味のチーズケーキ【消費】", 14, 0, 0x0000, 6, 255, 0x0084, 88, 128, 9, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 38, 38, 1, 0},
+    {_S("ショートケーキ"), "特製クリームと苺のショートケーキ【消費】", 11, 0, 0x0000, 6, 255, 0x0084, 88, 128, 10, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 39, 39, 1, 0},
+    {_S("いちごのタルト"), "採れたてのいちごのタルト【消費】", 12, 0, 0x0000, 6, 255, 0x0084, 88, 128, 11, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 40, 40, 1, 0},
+    {_S("パフェ"), "新鮮な生クリームがたっぷり", 16, 0, 0x0000, 6, 255, 0x0084, 88, 128, 12, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 41, 41, 1, 0},
+    {_S("プリンアラモード"), "極上のプリンアラモード【消費】", 17, 0, 0x0000, 6, 255, 0x0084, 88, 128, 13, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 42, 42, 1, 0},
+    {_S("アイスクリーム"), "３段重ねのアイスクリーム【消費】", 15, 0, 0x0000, 6, 255, 0x0084, 88, 128, 14, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 43, 43, 1, 0},
+    {_S("フランクフルト"), "地元の名品【消費】", 19, 0, 0x0000, 6, 255, 0x0084, 88, 128, 15, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 44, 44, 1, 0},
+    {_S("ハンバーガー"), "牛肉１００％【消費】", 20, 0, 0x0000, 6, 255, 0x0084, 88, 128, 16, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 45, 45, 1, 0},
+    {_S("ピザ"), "イタリアンクラストにダブルチーズ【消費】", 21, 0, 0x0000, 6, 255, 0x0084, 88, 128, 17, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 46, 46, 1, 0},
+    {_S("チーズ"), "うまいチーズ【消費】", 18, 0, 0x0000, 6, 255, 0x0084, 88, 128, 18, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 47, 47, 1, 0},
+    {_S("ハムエッグ"), "厳選された玉子を使用【消費】", 22, 0, 0x0000, 6, 255, 0x0084, 88, 128, 19, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 48, 48, 1, 0},
+    {_S("ミートオムレツ"), "まろやかな味のハーモニー【消費】", 23, 0, 0x0000, 6, 255, 0x0084, 88, 128, 20, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 49, 49, 1, 0},
+    {_S("モーニングセット"), "エッグトーストと牛乳【消費】", 24, 0, 0x0000, 6, 255, 0x0084, 88, 128, 21, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 50, 50, 1, 0},
+    {_S("Aランチ"), "お昼のサービスセット【消費】", 25, 0, 0x0000, 6, 255, 0x0084, 88, 128, 22, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 51, 51, 1, 0},
+    {_S("Bランチ"), "お昼のおすすめセット【消費】", 26, 0, 0x0000, 6, 255, 0x0084, 88, 128, 23, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 52, 52, 1, 0},
+    {_S("カレーライス"), "豊富にスパイスを煮込んだカレー【消費】", 28, 0, 0x0000, 6, 255, 0x0084, 88, 128, 24, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 53, 53, 1, 0},
+    {_S("ドリア"), "とろけるチーズがたっぷり【消費】", 29, 0, 0x0000, 6, 255, 0x0084, 88, 128, 25, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 54, 54, 1, 0},
+    {_S("スパゲティー"), "コシがあるのに歯切れが良い【消費】", 27, 0, 0x0000, 6, 255, 0x0084, 88, 128, 26, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 55, 55, 1, 0},
+    {_S("ワイン"), "豊潤な香りの１７２０年もの【消費】", 20, 0, 0x0000, 6, 255, 0x0084, 88, 128, 27, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 56, 56, 1, 0},
+    {_S("ビール"), "キレの良い喉ごし【消費】", 15, 0, 0x0000, 6, 255, 0x0084, 88, 128, 28, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 57, 57, 1, 0},
+    {_S("おさけ"), "水口の誰にでも合う日本酒の銘品【消費】", 10, 0, 0x0000, 6, 255, 0x0084, 88, 128, 29, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 58, 58, 1, 0},
+    {_S("なっとう"), "東洋の珍味【消費】", 32, 0, 0x0000, 6, 255, 0x0084, 88, 128, 30, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 59, 59, 1, 0},
+    {_S("ラーメン"), "特製""\x83\x5C""ースがうまさのきめて【消費】", 30, 0, 0x0000, 6, 255, 0x0084, 88, 128, 31, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 60, 60, 1, 0},
+    {_S("みそしる"), "新鮮なカニの入った味噌汁【消費】", 20, 0, 0x0000, 6, 255, 0x0084, 88, 128, 32, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 61, 61, 1, 0},
+    {_S("すし"), "握り寿司（大トロ、エビ、玉子）【消費】", 100, 0, 0x0000, 6, 255, 0x0084, 88, 128, 33, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 62, 62, 1, 0},
+    {_S("にくまん"), "挽肉の入った中華まんじゅう【消費】", 17, 0, 0x0000, 6, 255, 0x0084, 88, 128, 34, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 63, 63, 1, 0},
+    {_S("あんまん"), "あんこの入った中華まんじゅう【消費】", 12, 0, 0x0000, 6, 255, 0x0084, 88, 128, 35, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 64, 64, 1, 0},
+    {_S("てんしん"), "うまい点心は何個でも食える【消費】", 10, 0, 0x0000, 6, 255, 0x0084, 88, 128, 36, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 65, 65, 1, 0},
+    {_S("てんしんセット"), "点心の３点セット【消費】", 30, 0, 0x0000, 6, 255, 0x0084, 88, 128, 37, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 66, 66, 1, 0},
+    {_S("うまいにく"), "悪魔城特産の回復食料【消費】", 50, 0, 0x0000, 6, 255, 0x0084, 88, 128, 38, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 67, 67, 1, 0},
+    {_S("ビーフ"), "ビーフステーキ【消費】", 100, 0, 0x0000, 6, 255, 0x0084, 88, 128, 39, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 68, 68, 1, 0},
+    {_S("ターキー"), "七面鶏の肉【消費】", 80, 0, 0x0000, 6, 255, 0x0084, 88, 128, 40, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 69, 69, 1, 0},
+    {_S("くさいめし"), "賞味期限切れ【消費】", 40, 0, 0x0000, 6, 255, 0x0084, 88, 128, 41, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 2, 62, 1, 0},
+    {_S("うまいめし"), "フルコース（５つ星）【消費】", 75, 0, 0x0000, 6, 255, 0x0084, 88, 128, 42, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 3, 54, 1, 0},
+    {_S("おしょくじけん"), "お食事券　何が出るかわからない【消費】", 40, 0, 0x0000, 6, 255, 0x0084, 88, 128, 43, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 70, 70, 1, 0},
+    {_S("ニュートロンボム"), "全周囲の敵にダメージ【消費】", 30, 0, 0x0000, 7, 19, 0x0082, 88, 50, 0, 0, 256, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 130, 1, 71, 71, 1, 0},
+    {_S("おやじのいこう"), "親父の威光で全ての敵を攻撃【消費】", 40, 0, 0x0000, 7, 18, 0x0082, 88, 50, 0, 0, 256, 0, 1, 8, 0, 0x00000000, 0x00000000, 0, 4, 130, 1, 72, 72, 1, 0},
+    {_S("ペンタグラム"), "全周囲の敵にダメージ【消費】", 50, 0, 0x0000, 7, 17, 0x0082, 88, 50, 0, 0, 256, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 130, 1, 73, 73, 1, 0},
+    {_S("バットペンタグラム"), "全周囲の敵にダメージ【消費】", 99, 0, 0x0000, 7, 59, 0x0082, 88, 50, 1, 0, 256, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 130, 1, 74, 74, 1, 0},
+    {_S("しゅりけん"), "小型の手裏剣【消費】", 50, 0, 0x0040, 8, 15, 0x0082, 88, 50, 0, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 75, 75, 32, 0},
+    {_S("じゅうじしゅりけん"), "遠心力で威力をました手裏剣【消費】", 70, 0, 0x0040, 8, 15, 0x0081, 85, 50, 1, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 76, 76, 32, 0},
+    {_S("やぎゅうしゅりけん"), "上忍が使う磨かれた手裏剣【消費】", 90, 0, 0x0040, 8, 15, 0x0082, 88, 50, 2, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 77, 77, 32, 0},
+    {_S("炎のしゅりけん"), "火炎属性の忍術手裏剣【消費】", 100, 0, 0x8040, 8, 15, 0x0081, 85, 50, 3, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 78, 78, 32, 0},
+    {_S("ばくらい"), "火柱があがる【消費】", 60, 0, 0x8000, 8, 15, 0x0082, 88, 68, 0, 0, 256, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 79, 79, 1, 0},
+    {_S("ブワカナイフ"), "ブワカ族の使う投げナイフ【消費】", 30, 0, 0x0040, 8, 16, 0x0082, 85, 50, 0, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 80, 80, 1, 0},
+    {_S("ブーメラン"), "対空攻撃用の手投げ武器【消費】", 30, 0, 0x0000, 8, 16, 0x0081, 85, 62, 0, 0, 768, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 81, 81, 1, 0},
+    {_S("ジャベリン"), "手投げ槍【消費】", 50, 0, 0x0040, 8, 16, 0x0081, 85, 68, 0, 0, 512, 0, 1, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 82, 82, 1, 0},
+    {_S("ティルフィング"), "呪われている暗黒剣", -30, 0, 0x0940, 1, 0, 0x0001, 57, 48, 0, 18, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 83, 83, 1, 0},
+    {_S("ナマクラ"), "適当に造られた鈍い刀【両手持ち】", -5, 0, 0x0140, 5, 0, 0x0001, 57, 48, 0, 18, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 84, 84, 1, 0},
+    {_S("ナックル"), "拳にはめる鉛製拳鍔", 1, 0, 0x0000, 3, 1, 0x0000, 27, 48, 0, 5, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 85, 85, 64, 0},
+    {_S("グラディウス"), "古代ローマの剣", 3, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 86, 86, 1, 0},
+    {_S("シミター"), "新月刀。湾曲した片刃の剣", 5, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 87, 87, 1, 0},
+    {_S("カトラス"), "英国海軍の軍刀（カットラス）", 7, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 88, 88, 1, 0},
+    {_S("セイバー"), "騎兵用の軍刀（サーベル）", 9, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 89, 89, 1, 0},
+    {_S("ファルシオン"), "ノルマン人の使った幅広のえん月刀", 10, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 90, 90, 1, 0},
+    {_S("カッツバルゲル"), "傭兵が使ったシンプルな剣", 12, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 91, 91, 1, 0},
+    {_S("ベカトワ"), "飾り気のない質素な実戦的で扱いやすい刀剣", 14, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 92, 92, 1, 0},
+    {_S("ダマスカスソード"), "優れた素材の堅く歯こぼれしない刀剣", 17, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 93, 93, 1, 0},
+    {_S("ジャドプラーテ"), "貴族が狩猟用に用いた優雅な作りの刀剣", 20, -1, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 94, 94, 8, 0},
+    {_S("エストック"), "ドイツの両手突き剣【両手持ち】", 39, 0, 0x0040, 5, 51, 0x001B, 103, 48, 8, 18, 32768, 0, 0, 28, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 95, 95, 32, 0},
+    {_S("バスタードソード"), "柄が長い片手半剣", 20, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 96, 96, 1, 0},
+    {_S("ダイアナックル"), "ダイアモンドで強化した拳鍔", 21, 0, 0x0000, 3, 1, 0x0000, 27, 48, 1, 5, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 97, 97, 64, 0},
+    {_S("クレイモア"), "スコットランドの本式の両手剣【両手持ち】", 26, 0, 0x0040, 5, 51, 0x0006, 96, 48, 0, 15, 32957, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 98, 98, 1, 0},
+    {_S("ターワル"), "インドの曲刃剣（タルワール）", 22, 0, 0x0040, 1, 0, 0x0001, 57, 48, 0, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 99, 99, 1, 0},
+    {_S("カタナ"), "無名の刀匠による真打【両手持ち】", 32, 0, 0x0040, 5, 53, 0x0001, 57, 48, 0, 13, 32768, 196, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 100, 100, 32, 0},
+    {_S("フランベルジェ"), "火炎形刃の儀礼用両手剣【両手持ち】", 31, 0, 0x0040, 5, 51, 0x0006, 96, 48, 1, 15, 32958, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 101, 101, 1, 0},
+    {_S("カイザーナックル"), "神の怒りが宿る拳鍔", 26, 0, 0x0000, 3, 1, 0x0000, 27, 48, 2, 5, 32768, 193, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 102, 102, 64, 0},
+    {_S("ツヴァイハンダー"), "両手剣【両手持ち】", 38, 0, 0x0040, 5, 51, 0x0006, 96, 48, 2, 15, 32959, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 103, 103, 1, 0},
+    {_S("ヘルゲの剣"), "ヘルゲの剣", 24, 0, 0x0040, 1, 0, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 104, 104, 1, 0},
+    {_S("オルクリスト"), "森の精に鍛えられたといわれている剣", 26, 0, 0x0040, 1, 0, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 105, 105, 1, 0},
+    {_S("ハルパー"), "ハルパー", 28, 0, 0x0040, 1, 0, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 106, 106, 1, 0},
+    {_S("こくよう石の剣"), "黒曜石で作られた剣【両手持ち】", 42, 0, 0x0040, 5, 51, 0x0006, 96, 48, 3, 13, 32960, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 107, 107, 1, 0},
+    {_S("名刀グラム"), "名刀グラム", 30, 0, 0x0040, 1, 0, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 108, 108, 1, 0},
+    {_S("ジュエルソード"), "敵を宝石に変える", 15, 0, 0x0040, 0, 37, 0x000F, 161, 48, 1, 13, 32768, 205, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 133, 109, 109, 1, 0},
+    {_S("ホーリーバスター"), "聖なる者に効く暗黒剣", 25, 0, 0x0840, 1, 0, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 110, 110, 1, 0},
+    {_S("アグニの炎剣"), "火炎属性の剣", 25, 0, 0x8040, 1, 49, 0x0001, 57, 48, 0, 13, 32768, 186, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 111, 111, 1, 0},
+    {_S("インドラの雷剣"), "雷撃属性の剣", 25, 0, 0x4040, 1, 48, 0x0001, 57, 48, 0, 13, 32768, 185, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 112, 112, 1, 0},
+    {_S("ラハブの氷剣"), "氷結属性の剣", 25, 0, 0x2040, 1, 50, 0x0001, 57, 48, 0, 13, 32768, 188, 0, 21, 0, 0x00000000, 0x00000000, 0, 8, 2, 129, 113, 113, 1, 0},
+    {_S("ミリカンの石刀"), "敵の停止時間が長い、石化してくれる敵も、、、", 1, 0, 0x0200, 0, 37, 0x000F, 161, 48, 3, 8, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 64, 2, 129, 114, 114, 1, 0},
+    {_S("ホーリーソード"), "アンデッドに効く聖剣", 26, 0, 0x1040, 1, 55, 0x0001, 57, 48, 0, 13, 32768, 203, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 115, 115, 1, 0},
+    {_S("フルンチング"), "毒の追加ダメージ", 32, 2, 0x00C0, 1, 0, 0x0201, 57, 48, 5, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 116, 116, 1, 0},
+    {_S("レーヴァテイン"), "強力な火炎属性の剣", 33, 1, 0x8040, 1, 49, 0x0001, 57, 48, 2, 13, 32768, 187, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 117, 117, 1, 0},
+    {_S("グラムドリング"), "かつて高名な魔術師が携えていた名剣", 35, 2, 0x0040, 1, 0, 0x0001, 57, 48, 2, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 118, 118, 1, 0},
+    {_S("ヘブンズソード"), "空中を舞い敵を切り裂く魔封剣", 21, -2, 0x0040, 2, 13, 0x0081, 85, 50, 0, 0, 256, 0, 0, 20, 0, 0x00000001, 0x00000001, 0, 4, 2, 129, 119, 119, 32, 0},
+    {_S("オリハルコン"), "熱き魂を秘めたナックル", 38, 5, 0x0000, 3, 1, 0x0000, 27, 48, 3, 5, 32962, 195, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 120, 120, 64, 0},
+    {_S("ダインスレイフ"), "血を吸ってパワーアップする魔剣", 25, 1, 0x0040, 1, 54, 0x0001, 57, 48, 0, 13, 32768, 202, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 121, 121, 1, 0},
+    {_S("モーンブレード"), "敵の生命力を持ち主に与える魔剣", 36, 1, 0x0040, 1, 0, 0x0101, 57, 48, 6, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 130, 122, 122, 1, 0},
+    {_S("アルカードソード"), "母の形見の剣", 42, 0, 0x0040, 1, 43, 0x0001, 57, 48, 0, 10, 32768, 204, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 123, 123, 1, 0},
+    {_S("レムリアソード"), "亡霊祭の剣、盾の魔力を引き出す", 39, 2, 0x0040, 1, 57, 0x0001, 57, 48, 3, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00007FFE, 0, 4, 2, 129, 124, 124, 1, 0},
+    {_S("バルザイのえん月刀"), "プレイ時間に応じて威力が上昇する", 0, 0, 0x0040, 1, 0, 0x0001, 57, 48, 3, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 125, 125, 1, 0},
+    {_S("つかい魔の剣"), "持ち主を選ぶレベルアップする魔剣", 0, 0, 0x0040, 1, 43, 0x0101, 57, 48, 2, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 126, 126, 1, 0},
+    {_S("グレートソード"), "物理ダメージが大きい【両手持ち】", 65, 0, 0x0040, 5, 51, 0x0006, 96, 48, 0, 16, 32768, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 127, 127, 1, 0},
+    {_S("メイス"), "金属製こん棒", 9, 0, 0x0000, 4, 2, 0x0002, 64, 48, 0, 11, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 128, 128, 1, 0},
+    {_S("モーニングスター"), "明けの明星の名を冠したこん棒", 11, 0, 0x0000, 4, 3, 0x0002, 64, 48, 0, 11, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 129, 129, 1, 0},
+    {_S("ホーリーロッド"), "聖なる力を秘めたロッド", 12, 0, 0x1000, 4, 4, 0x0003, 71, 48, 0, 11, 32945, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 130, 130, 1, 0},
+    {_S("スターフレイル"), "星の力がフレイル状に連なるロッド", 15, 0, 0x0000, 4, 5, 0x0003, 71, 48, 0, 11, 32946, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 131, 131, 1, 0},
+    {_S("ムーンロッド"), "月の形を模したロッド", 17, 0, 0x0000, 4, 6, 0x0002, 64, 48, 0, 11, 32768, 179, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 132, 132, 1, 0},
+    {_S("チャクラム"), "薄く鋭利な鋼鉄の手投げ武器　戦輪", 15, 0, 0x0040, 2, 12, 0x0082, 88, 56, 0, 0, 512, 0, 0, 30, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 133, 133, 32, 0},
+    {_S("ほのおのブーメラン"), "火炎属性の対空用手投げ武器【消費】", 50, 0, 0x8000, 8, 16, 0x0081, 85, 62, 1, 0, 768, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 134, 134, 1, 0},
+    {_S("てっきゅう"), "ヘビーアーマーが使用していた鉄球【消費】", 60, 0, 0x0000, 8, 16, 0x0081, 85, 63, 0, 0, 1024, 0, 1, 30, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 135, 135, 1, 0},
+    {_S("はやわざのたんとう"), "斜め前方にすばやい攻撃", 6, 0, 0x0040, 0, 46, 0x0086, 202, 50, 0, 0, 256, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 18, 18, 32, 0},
+    {_S("ブルーナックル"), "アッパーカット専用のナックル", 8, 0, 0x0000, 3, 46, 0x0086, 202, 56, 0, 0, 256, 0, 0, 18, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 85, 85, 64, 0},
+    {_S("ダイナマイト"), "大爆発！たまには不発も、、、【消費】", 100, 0, 0x8000, 8, 47, 0x0082, 88, 50, 0, 0, 256, 0, 1, 16, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 136, 79, 1, 0},
+    {_S("びぜんおさふね"), "刀匠長船の手による真打【両手持ち】", 50, -5, 0x0040, 5, 53, 0x0001, 57, 48, 2, 13, 32967, 200, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 100, 100, 32, 0},
+    {_S("まさむね"), "刀匠正宗の手による真打【両手持ち】", 90, -15, 0x0040, 5, 53, 0x0001, 57, 48, 4, 13, 32768, 201, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 100, 100, 64, 0},
+    {_S("ようとうむらまさ"), "血に飢えた妖刀、呪われている【両手持ち】", -5, -2, 0x0140, 5, 56, 0x001D, 57, 48, 0, 13, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 26, 26, 1, 0},
+    {_S("ハートリペア"), "ハート５００個回復【消費】", 0, 0, 0x0000, 10, 44, 0x0083, 182, 50, 0, 0, 0, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 142, 142, 1, 0},
+    {_S("ヘブンズルーン"), "刃に魔呪を刻んだ飛剣", 40, -4, 0x0040, 2, 14, 0x0081, 85, 50, 0, 0, 256, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 143, 143, 32, 0},
+    {_S("セィラム"), "毒から回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 0, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 144, 144, 1, 0},
+    {_S("アンカース"), "呪いから回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 1, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 145, 144, 1, 0},
+    {_S("せいめいのれいやく"), "死んだとき、蘇る（自分では使えない）【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 2, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 154, 149, 1, 0},
+    {_S("ハンマー"), "石化から回復（自分では使えない）【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 3, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 155, 149, 1, 0},
+    {_S("ちからのくすり"), "一時的に力を上げる【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 4, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 149, 149, 1, 0},
+    {_S("ラッキードラッグ"), "一時的に運を上げる【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 5, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 150, 151, 1, 0},
+    {_S("ちりょくのくすり"), "一時的に知力を上げる【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 6, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 150, 149, 1, 0},
+    {_S("アタックポーション"), "一時的に攻撃力を上げる【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 7, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 150, 147, 1, 0},
+    {_S("シールドポーション"), "一時的に防御力を上げる【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 8, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 150, 150, 1, 0},
+    {_S("レジストファイア"), "一時的に対火炎属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 9, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 148, 1, 0},
+    {_S("レジストサンダー"), "一時的に対雷撃属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 10, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 150, 1, 0},
+    {_S("レジストアイス"), "一時的に対氷結属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 11, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 152, 1, 0},
+    {_S("レジストストーン"), "一時的に対石化属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 12, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 145, 1, 0},
+    {_S("レジストホーリー"), "一時的に対聖属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 14, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 149, 1, 0},
+    {_S("レジストダーク"), "一時的に対暗黒属性にする【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 15, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 148, 154, 1, 0},
+    {_S("ポーション"), "ＨＰを少し回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 16, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 146, 151, 1, 0},
+    {_S("ハイポーション"), "ＨＰを中ぐらい回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 17, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 147, 151, 1, 0},
+    {_S("エクスポーション"), "ＨＰを全回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 18, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 151, 151, 1, 0},
+    {_S("マナプリズム"), "ＭＰを全回復【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 19, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 152, 153, 1, 0},
+    {_S("ソニックブレード"), "真空の刃が敵を切り裂く", 25, 0, 0x0040, 1, 12, 0x0081, 85, 50, 1, 0, 768, 0, 0, 14, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 119, 119, 1, 0},
+    {_S("ヴァルマンウェ"), "無数の真空の刃が敵を切り刻む", 36, 1, 0x0040, 1, 12, 0x0081, 85, 52, 2, 0, 768, 0, 0, 14, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 119, 119, 1, 0},
+    {_S("やすつな"), "最古の刀工安綱が鍛えた銘刀【両手持ち】", 60, -10, 0x0040, 5, 53, 0x0081, 85, 68, 0, 0, 256, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 100, 100, 64, 0},
+    {_S("マジカルチケット"), "爺が恋しいときに使うべし【消費】", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 20, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 70, 70, 1, 0},
+    {_S("サインいりしたじき"), "ゲットだぜ！！", 0, 0, 0x0000, 10, 255, 0x0083, 182, 55, 20, 0, 256, 0, 1, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 156, 141, 1, 0},
+    {_S("アルカートシールド"), "以前から愛用している盾に似ている", 1, 0, 0x0000, 9, 58, 0x0080, 91, 48, 1, 0, 32768, 0, 0, 6, 0, 0x00002000, 0x00000000, 0, 0, 4, 1, 4, 4, 1, 0},
+    {_S("アルカートソード"), "母の形見の剣に似ている", 2, 0, 0x0040, 1, 43, 0x0001, 57, 48, 1, 10, 32768, 0, 0, 21, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 123, 123, 16, 0},
+    {_S("アルカードスピア"), "かつて愛用していた妖槍【両手持ち】", 28, 0, 0x0040, 5, 60, 0x0006, 96, 48, 0, 15, 32992, 225, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 210, 138, 1, 0},
+    {_S("ウッドランドシールド"), "樹霊が宿る盾（バラの香り）", 1, 3, 0x0000, 9, 61, 0x0080, 91, 48, 0, 0, 32768, 0, 0, 6, 0, 0x00004000, 0x00000000, 0, 0, 4, 1, 233, 211, 1, 0},
+    {_S("アストラルダガー"), "不思議な異次元の短剣", 12, 0, 0x0040, 0, 62, 0x0009, 119, 48, 0, 5, 32768, 227, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 212, 140, 32, 0},
+    {_S("ダブルブレードアックス"), "２枚刃で良く切れる【両手持ち】", 42, 0, 0x0040, 5, 63, 0x0006, 96, 48, 0, 13, 32768, 0, 0, 15, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 214, 136, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 1, 0, 0x0000, 57, 0, 0, 0, 0, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 27, 0, 0x1000, 4, 4, 0x0004, 78, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 32, 0},
+    {_S(""), "", 35, 0, 0x0000, 4, 5, 0x0004, 78, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 32, 0},
+    {_S(""), "", 35, 0, 0x0000, 4, 6, 0x0005, 94, 72, 0, 0, 256, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 4, 2, 1, 0, 0, 16, 0},
+    {_S(""), "", 10, 0, 0x0000, 4, 7, 0x0002, 64, 48, 1, 13, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 5, 0, 0x0040, 0, 31, 0x0008, 112, 48, 1, 8, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 15, 0, 0x0040, 0, 32, 0x000A, 126, 48, 1, 8, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 64, 0},
+    {_S(""), "", 20, 0, 0x0040, 1, 34, 0x0081, 85, 56, 0, 0, 256, 0, 0, 30, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 40, 0, 0x0000, 1, 30, 0x0005, 178, 56, 0, 0, 2048, 0, 0, 10, 0, 0x00000000, 0x00000000, 5, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 40, 0, 0x4040, 1, 48, 0x0013, 57, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 40, 0, 0x8040, 1, 49, 0x0014, 57, 48, 1, 13, 32768, 0, 0, 30, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 60, 0, 0x8040, 1, 49, 0x0015, 57, 48, 3, 13, 32768, 0, 0, 30, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 40, 0, 0x2040, 1, 50, 0x0016, 57, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 8, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 44, 0, 0x0040, 1, 51, 0x0017, 103, 54, 0, 13, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 56, 0, 0x0040, 1, 51, 0x0017, 103, 54, 1, 13, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 60, 0, 0x0040, 1, 51, 0x0017, 103, 54, 2, 13, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 68, 0, 0x0040, 1, 51, 0x0017, 103, 54, 3, 13, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 40, 0, 0x0000, 3, 1, 0x0018, 27, 54, 0, 5, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 64, 0},
+    {_S(""), "", 70, 0, 0x0000, 3, 1, 0x0018, 27, 54, 1, 5, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 64, 0},
+    {_S(""), "", 40, 0, 0x8000, 2, 1, 0x0019, 27, 48, 4, 7, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 1, 0, 0, 64, 0},
+    {_S(""), "", 50, 0, 0x0040, 5, 53, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 40, 0, 0x0040, 0, 39, 0x001A, 168, 48, 1, 13, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 15, 0, 0x0040, 0, 38, 0x001A, 168, 48, 1, 13, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 100, 0, 0x0040, 5, 53, 0x0001, 57, 48, 3, 13, 32768, 0, 0, 14, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 130, 0, 0x0040, 5, 53, 0x0087, 206, 60, 0, 0, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 200, 0, 0x0040, 5, 53, 0x0087, 206, 60, 1, 0, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 40, 0, 0x0040, 1, 54, 0x0001, 57, 48, 1, 13, 32768, 0, 0, 10, 0, 0x00000000, 0x00000000, 0, 4, 2, 129, 0, 0, 1, 0},
+    {_S(""), "", 35, 0, 0x1040, 1, 55, 0x001C, 57, 48, 1, 13, 32768, 0, 0, 14, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 50, 0, 0x0040, 1, 43, 0x0087, 207, 54, 0, 0, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 5, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 15, 0, 0x0040, 0, 37, 0x000F, 161, 48, 2, 13, 32768, 0, 0, 12, 0, 0x00000000, 0x00000000, 5, 4, 2, 133, 0, 0, 1, 0},
+    {_S(""), "", 40, 0, 0x2040, 1, 50, 0x0016, 57, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 8, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 40, 0, 0x2040, 1, 50, 0x0016, 57, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 8, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 40, 0, 0x2040, 1, 50, 0x0016, 57, 48, 1, 13, 32768, 0, 0, 20, 0, 0x00000000, 0x00000000, 5, 8, 2, 129, 0, 0, 16, 0},
+    {_S(""), "", 50, 0, 0x0040, 2, 13, 0x0005, 94, 62, 0, 0, 256, 0, 0, 12, 0, 0x00000001, 0x00000000, 5, 4, 2, 129, 0, 0, 64, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 8, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000002, 0x00000000, 10, 4, 0, 1, 0, 0, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 9, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000004, 0x00000000, 10, 4, 0, 1, 0, 0, 1, 0},
+    {_S(""), "", 40, 0, 0x0000, 1, 10, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000008, 0x00000000, 20, 4, 2, 1, 0, 0, 64, 0},
+    {_S(""), "", 20, 0, 0x0000, 1, 11, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000010, 0x00000000, 20, 4, 2, 1, 0, 0, 64, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 23, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000020, 0x00000000, 10, 4, 0, 1, 0, 0, 1, 0},
+    {_S(""), "", 30, 0, 0x0000, 7, 24, 0x0005, 190, 86, 0, 0, 256, 0, 0, 80, 0, 0x00000040, 0x00000000, 30, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 25, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000080, 0x00000000, 40, 4, 0, 1, 0, 0, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 26, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000100, 0x00000000, 10, 4, 0, 1, 0, 0, 1, 0},
+    {_S(""), "", 25, 0, 0x0000, 1, 27, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00000200, 0x00000000, 30, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 100, 0, 0x0000, 1, 28, 0x0005, 190, 86, 0, 0, 256, 0, 0, 20, 0, 0x00000400, 0x00000000, 40, 4, 2, 1, 0, 0, 32, 0},
+    {_S(""), "", 70, 0, 0x8000, 1, 29, 0x0005, 190, 86, 0, 0, 256, 0, 0, 18, 0, 0x00000800, 0x00000000, 50, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 30, 0, 0x8000, 10, 52, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00001000, 0x00000000, 20, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 58, 0x0005, 190, 86, 0, 0, 256, 0, 0, 10, 0, 0x00002000, 0x00000000, 10, 4, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 0, 0, 0x0000, 10, 45, 0x0083, 182, 55, 14, 0, 256, 0, 0, 0, 0, 0x00000000, 0x00000000, 0, 0, 0, 1, 157, 147, 1, 0},
+    {_S(""), "", 50, 0, 0x0040, 1, 60, 0x0020, 224, 54, 1, 24, 32768, 0, 0, 22, 0, 0x00000000, 0x00000000, 7, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 20, 0, 0x0040, 1, 60, 0x0021, 96, 60, 2, 15, 32768, 0, 0, 8, 0, 0x00000000, 0x00000000, 7, 4, 2, 129, 0, 0, 32, 0},
+    {_S(""), "", 30, 0, 0x0000, 1, 61, 0x0005, 190, 86, 0, 0, 256, 0, 0, 30, 0, 0x00004000, 0x00000000, 30, 200, 2, 1, 0, 0, 1, 0},
+    {_S(""), "", 10, 0, 0x0040, 0, 62, 0x0005, 190, 90, 0, 0, 256, 0, 0, 10, 0, 0x00000020, 0x00000000, 4, 4, 2, 1, 0, 0, 1, 0},
+};
+
+Accessory g_AccessoryDefs[] = {
+    {_S("----"), "装備解除", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 184, 184, 1, 0},
+    {_S("ぬののふく"), "着心地のよい衣服", 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 185, 185, 1, 0},
+    {_S("かわのむねあて"), "なめし皮製の胸あて", 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 186, 186, 1, 0},
+    {_S("どうのむねあて"), "銅の胸あて", 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 187, 187, 1, 0},
+    {_S("てつのむねあて"), "鉄の胸あて", 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 188, 188, 1, 0},
+    {_S("はがねのむねあて"), "鋼の胸あて", 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 189, 189, 1, 0},
+    {_S("シルバーチェスト"), "銀の胸あて", 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 190, 190, 1, 0},
+    {_S("ゴールドチェスト"), "金の胸あて", 0, 8, 0, 0, 0, 0, 0, 0, 0, 0, 191, 191, 1, 0},
+    {_S("プラチナチェスト"), "白金の胸あて", 0, 10, 0, 0, 0, 0, 0, 0, 0, 0, 192, 192, 1, 0},
+    {_S("ダイアブレスト"), "ダイアの胸あて", 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 193, 193, 1, 0},
+    {_S("炎のむねあて"), "火炎属性攻撃に強い", 0, 4, 0, 0, 0, 0, 0, 32768, 0, 0, 194, 194, 1, 0},
+    {_S("雷のむねあて"), "雷撃属性攻撃に強い", 0, 8, 0, 0, 0, 0, 0, 16384, 0, 0, 195, 195, 1, 0},
+    {_S("氷のむねあて"), "氷結属性攻撃に強い", 0, 8, 0, 0, 0, 0, 0, 8192, 0, 0, 196, 196, 1, 0},
+    {_S("ミラーキュイラス"), "石化攻撃に耐える鏡のように磨かれた胸あて", 0, 3, 0, 0, 0, 0, 0, 0, 512, 0, 197, 197, 1, 0},
+    {_S("スパイクブレイカ"), "刺地形を壊す鎧", 0, 12, 0, 0, 0, 0, 0, 0, 16, 0, 198, 198, 1, 0},
+    {_S("アルカードメイル"), "火炎、雷撃、氷結に強い", 0, 20, 0, 0, 0, 0, 0, 57344, 0, 0, 199, 199, 1, 0},
+    {_S("邪なむねあて"), "暗黒属性でダメージを受けない", 0, 8, 0, 0, 0, 0, 4096, 0, 2048, 0, 200, 200, 1, 0},
+    {_S("ヒーリングメール"), "歩くたびにＨＰ回復", 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 201, 201, 1, 0},
+    {_S("聖なるむねあて"), "聖なる属性のダメージを軽減", 0, 11, 0, 0, 0, 0, 0, 4096, 0, 0, 202, 202, 1, 0},
+    {_S("ウォークマスター"), "うめたマップ数に応じて防御力が上昇", 0, 0, 0, 0, 251, 0, 0, 0, 0, 0, 203, 203, 1, 0},
+    {_S("光のむねあて"), "サブウェポンの攻撃強化", 0, 12, 0, 0, 1, 0, 0, 0, 0, 0, 204, 204, 1, 0},
+    {_S("やみのむねあて"), "魔術の攻撃強化", 0, 15, 0, 0, 1, 0, 0, 0, 0, 0, 205, 205, 1, 0},
+    {_S("いかりのむねあて"), "ダメージを受けると一定時間防御力アップ", 1, 13, 0, 0, 0, 0, 0, 0, 0, 0, 206, 206, 1, 0},
+    {_S("ドラキュラのふく"), "貴族の服", 2, 14, 1, 1, 1, 0, 0, 0, 0, 0, 207, 207, 1, 0},
+    {_S("ガーブオブゴッド"), "全てにおいて最強", 0, 25, 0, 2, 0, 0, 0, 0, 0, 0, 208, 208, 1, 0},
+    {_S("アックスアーマーのよろい"), "アックスアーマーになる", 2, 15, 0, 0, 0, 0, 0, 0, 512, 0, 209, 209, 1, 0},
+    {_S("----"), "装備解除", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 159, 159, 0, 0},
+    {_S("メガネ"), "カッコイイ眼鏡", 0, 1, 0, 0, 255, 0, 0, 0, 0, 0, 160, 160, 0, 0},
+    {_S("ダンスのマスク"), "舞踏会で使われる仮面", 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 161, 161, 0, 0},
+    {_S("バンダナ"), "頭部急所を隠すことで防御力上昇", 0, 2, 0, 0, 1, 0, 0, 0, 0, 0, 162, 162, 0, 0},
+    {_S("フェルトのぼうし"), "フェルトの帽子", 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 163, 163, 0, 0},
+    {_S("ビロードのぼうし"), "ビロードの帽子", 0, 1, 0, 0, 2, 0, 0, 0, 0, 0, 164, 164, 0, 0},
+    {_S("ゴーグル"), "眼球を保護し心理的余裕をつくる", 0, 1, 0, 0, 4, 0, 0, 0, 0, 0, 165, 165, 0, 0},
+    {_S("かわのぼうし"), "なめし皮の帽子", 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 166, 166, 0, 0},
+    {_S("聖なるめがね"), "魔呪で呪われた者を見破る", 0, 1, 0, 0, 5, 0, 0, 0, 0, 0, 167, 167, 0, 0},
+    {_S("たび人ぼう"), "旅人の基本装備である帽子", 0, 3, 0, 0, 1, 0, 0, 0, 0, 0, 168, 168, 0, 0},
+    {_S("石かめん"), "古代の祭儀に使われた仮面。魔法防御あり", 0, 2, 0, 0, 4, 0, 0, 0, 0, 0, 169, 169, 0, 0},
+    {_S("サークレット"), "サークレット", 0, 0, 0, 0, 5, 0, 0, 0, 0, 0, 170, 170, 0, 0},
+    {_S("ぎんのサークレット"), "聖なるダメージでＨＰ回復", 0, 0, 0, 0, 6, 0, 0, 0, 0, 4096, 171, 171, 0, 0},
+    {_S("こう玉石のサークレット"), "火炎ダメージでＨＰ回復", 0, 0, 0, 0, 10, 0, 0, 0, 0, 32768, 172, 147, 0, 0},
+    {_S("せい玉石のサークレット"), "氷結ダメージでＨＰ回復", 0, 0, 0, 0, 10, 0, 0, 0, 0, 8192, 172, 151, 0, 0},
+    {_S("おう玉石のサークレット"), "毒のダメージでＨＰ回復", 0, 0, 0, 0, 10, 0, 0, 0, 0, 128, 172, 149, 0, 0},
+    {_S("りょくちゅう石のサークレット"), "雷撃ダメージでＨＰ回復", 0, 0, 0, 0, 10, 0, 0, 0, 0, 16384, 173, 150, 0, 0},
+    {_S("ねこめ石のサークレット"), "猫のダメージでＨＰ大幅に回復", 0, 0, 0, 0, 8, 0, 0, 0, 0, 0, 173, 149, 0, 0},
+    {_S("さんごのサークレット"), "呪いの無効化", 0, 0, 0, 0, 7, 0, 0, 0, 256, 0, 170, 172, 0, 0},
+    {_S("エギルのかぶと"), "敵に恐怖を与え、防御力ダウン", 0, 3, 4, 0, 0, 0, 0, 0, 0, 0, 178, 178, 0, 0},
+    {_S("シルバーティアラ"), "知力アップ", 0, 0, 0, 0, 12, 0, 0, 0, 0, 0, 179, 179, 0, 0},
+    {_S("つばひろのたび人ぼう"), "高い防御力を持つ", 0, 5, 1, 1, 7, 0, 0, 0, 0, 0, 180, 180, 0, 0},
+    {_S("----"), "装備解除", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 215, 215, 2, 0},
+    {_S("ぬののマント"), "丈夫な布で体を守る。小汚なく見えるのが欠点", 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 216, 216, 2, 0},
+    {_S("リバーシブルマント"), "裏返すことで二色が選べるマント", 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 217, 217, 2, 0},
+    {_S("エルフのローブ"), "森で身を隠すのに最適な緑色のマント", 0, 0, 0, 5, 0, 0, 0, 0, 0, 0, 218, 218, 2, 0},
+    {_S("にじいろのローブ"), "時とともに色の変化するマント", 0, 0, 0, 7, 0, 0, 0, 0, 0, 0, 236, 236, 2, 0},
+    {_S("グレアリングクロス"), "闇を切り裂く白銀の輝き", 0, 0, 0, 6, 0, 0, 0, 0, 2048, 0, 231, 219, 2, 0},
+    {_S("ブルーマント"), "青く美しいマント", 0, 0, 0, 8, 0, 0, 0, 0, 0, 0, 220, 220, 2, 0},
+    {_S("血ぞめのマント"), "装着者のダメージをハートに変換する", 0, 0, 0, 4, 0, 255, 0, 0, 0, 0, 221, 221, 2, 0},
+    {_S("オーダーメイド"), "好きな色を選べる【システムメニュー】", 0, 0, 0, 9, 0, 0, 0, 0, 0, 0, 222, 222, 2, 0},
+    {_S("しっこくのマント"), "吸血鬼用の黒いマント", 0, 0, 0, 10, 0, 0, 0, 0, 0, 0, 223, 223, 2, 0},
+    {_S("----"), "装備解除", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 215, 215, 3, 0},
+    {_S("ムーンストーン"), "１８〜６時にステータスアップ", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 224, 224, 3, 0},
+    {_S("サンストーン"), "６〜１８時にステータスアップ", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 225, 225, 3, 0},
+    {_S("ブラッドストーン"), "血で回復倍増", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 226, 226, 3, 0},
+    {_S("STAUROLITE"), "\x8F\x5C""字架の威力アップ", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 227, 227, 3, 0},
+    {_S("パールの指輪"), "聖なる力を秘めた指輪", 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 228, 228, 3, 0},
+    {_S("ジルコン"), "高価にみえる指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 150, 3, 0},
+    {_S("アクアマリン"), "ちょっと高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 151, 3, 0},
+    {_S("ターコイズ"), "ちょっと高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 153, 3, 0},
+    {_S("オニキス"), "高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 154, 3, 0},
+    {_S("ガーネット"), "高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 148, 3, 0},
+    {_S("オパール"), "高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 230, 149, 3, 0},
+    {_S("ダイアモンド"), "とても高価な指輪【売却可】", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 229, 229, 3, 0},
+    {_S("ラピス・ラズリ"), "幸せを呼ぶ宝石", 0, 0, 0, 0, 0, 20, 0, 0, 0, 0, 230, 145, 3, 0},
+    {_S("ラウリンの指輪"), "２４人力の力を出せる指輪", 0, -24, 24, 0, 0, 0, 0, 0, 0, 0, 237, 237, 3, 0},
+    {_S("きんの指輪"), "ゴート語の刻印「対…る我……大………掲げ…」", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 239, 239, 3, 0},
+    {_S("ぎんの指輪"), "ゴート語の刻印「…な……等を…時計に……よ」", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 238, 238, 3, 0},
+    {_S("どくろの指輪"), "黄金の『持つものは世界を統べる』という指輪", 10, 10, 30, 10, 30, 5, 0, 0, 0, 0, 240, 240, 3, 0},
+    {_S("かしかーど"), "「夜曲」の歌詞カード", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 138, 74, 3, 0},
+    {_S("レアリング"), "稀少アイテム出現確率のアップ", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 241, 149, 3, 0},
+    {_S("ペンダント"), "魔力の回復が速まる", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 242, 242, 3, 0},
+    {_S("ハートのブローチ"), "ハートの消費量が減る", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 243, 243, 3, 0},
+    {_S("ネックレス"), "防御力＋５", 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 244, 244, 3, 0},
+    {_S("ガントレット"), "攻撃力＋５", 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 245, 245, 3, 0},
+    {_S("しゅくふくのアンク"), "ポーションなどの回復力アップ", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 246, 246, 3, 0},
+    {_S("マスターリング"), "\x94\x5C""力値を大幅に上げる", 1, 1, 7, 7, 7, 2, 0, 0, 0, 0, 247, 247, 3, 0},
+    {_S("メダル"), "防御力＋１　攻撃力＋１", 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 248, 248, 3, 0},
+    {_S("タリスマン"), "身代わりになってダメージを無効化することも", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 249, 249, 3, 0},
+    {_S("デュプリケーター"), "【消費】アイテムが減らなくなる", 0, -20, 0, 246, 246, 246, 0, 0, 0, 0, 250, 250, 3, 0},
+    {_S("キングストーン"), "力＋１０のバックル", 0, 0, 10, 0, 0, 0, 0, 0, 0, 0, 251, 251, 3, 0},
+    {_S("メギンギョルズ"), "力＋１５のバックル", 0, 0, 15, 0, 0, 0, 0, 0, 0, 0, 252, 252, 3, 0},
+    {_S("ブリーシンガメン"), "防御力＋１５の首飾り", 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 253, 253, 3, 0},
+    {_S("シークレットブーツ"), "誰にも気付かれずに背が伸びる！", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 254, 254, 3, 0},
+    {_S("アルカートメイル"), "火炎、雷撃、氷結に強いかもしれない", 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 199, 199, 1, 0},
+};
+
+const char* g_MenuStr[] = {
+    _S("ALUCARD"),
+    _S("STR"),
+    _S("CON"),
+    _S("INT"),
+    _S("LCK"),
+    _S("EXP"),
+    _S("NEXT"),
+    _S("GOLD"),
+    _S("LEVEL"),
+    _S("TIME"),
+    _S("ROOMS"),
+    _S("BEAT"),
+    _S("HP"),
+    _S("MP"),
+    _S("HEART"),
+    _S("STATUS"),
+    _S("ボタンせってい"),
+    _S("マントうらおもて"),
+    _S("マントちゃくしょく"),
+    _S("・・・・・"),
+    _S("おもて"),
+    _S("うら"),
+    _S("みぎ手"),
+    _S("ジャンプ"),
+    _S("ひだり手"),
+    _S("きり"),
+    _S("こうもり"),
+    _S("おおかみ"),
+    _S("Bダッシュ"),
+    _S("ホルダー"),
+    _S("ノーマル"),
+    _S("リバース"),
+    _S("サウンド"),
+    _S("ステレオ"),
+    _S("モノラル"),
+    _S("ウィンドウカラー"),
+    _S("タイムアタック"),
+    _S("GOOD"),
+    _S("POISON"),
+    _S("STONE"),
+    _S("CURSE"),
+    _S("-----------------"),
+    _S("ALUCART"),
+    _S("れべる"),
+    _S("けいけんち"),
+    _S("あるかーと"),
+    _S("コウモリ"),
+    _S("ゴースト"),
+    _S("妖精"),
+    _S("子悪魔"),
+    _S("剣魔"),
+    _S("ようせい"),
+    _S("はなあくま"),
+    _S("リヒターがドラキュラをたおす"),
+    _S("オルロックをたおす"),
+    _S("ドッペルゲンガーLV10をたおす"),
+    _S("レギオンをたおす"),
+    _S("ミノタウロスとウェアウルフをたおす"),
+    _S("スキュラをたおす"),
+    _S("ベリガンとギャイボンをたおす"),
+    _S("ヒポグリフをたおす"),
+    _S("ベルゼブブをたおす"),
+    _S("サキュバスをたおす"),
+    _S("マルファスをたおす"),
+    _S("ラルフとグラントとサイファをたおす"),
+    _S("デスをたおす"),
+    _S("ケルベロスをたおす"),
+    _S("リヒター・ベルモンドをたすける"),
+    _S("メディウサをたおす"),
+    _S("フランケンをたおす"),
+    _S("レッサーデーモンをたおす"),
+    _S("ドッペルゲンガーLV40をたおす"),
+    _S("ミイラ男をたおす"),
+    _S("ビッグこうもりをたおす"),
+    _S("ガラモスをたおす"),
+    _S("スケルトン・リーダーをたおす"),
+    _S("ラストセーブポイント"),
+    _S("デスにであう"),
+    _S("マリアに聖なるめがねをもらう"),
+    _S("としょかんのあるじとさいかいする"),
+    _S("はじめてマリアにあう"),
+    _S(""),
+    _S(""),
+    _S(""),
+    _S(""),
+    _S(""),
+};
+
+SpellDef g_SpellDefs[] = {
+    {"ダークメタモルフォーゼ", _S("←↖↑↗→"), "血をあびることにより体力回復", 10, 0, 0, 0, 1, 0, 0, 0, 0},
+    {"サモンスピリット", _S("←→↑↓"), "獰猛なスピリットを召喚", 5, 20, 8, 2, 1, 0, 0, 15, 0},
+    {"ヘルファイア", _S("↑↓↘→"), "亜空間移動した後、火炎攻撃を行う", 15, 20, 8, 2, 1, 0, 32768, 30, 0},
+    {"テトラスピリット", _S("↑ため↗→↘↓"), "スピリットを４体召喚", 20, 20, 8, 2, 1, 0, 0, 15, 0},
+    {"アクセルチャージ", _S("↓↘→"), "狼時に強力な体当たり攻撃", 10, 40, 10, 2, 129, 64, 64, 30, 0},
+    {"\x83\x5C""ウルスチール", _S("←→↘↓↙←→"), "付近の敵性体から体力を奪う", 50, 10, 8, 130, 2, 0, 0, 15, 0},
+    {"ウィングスマッシュ", _S("↑↖←↙↓↘→"), "コウモリ時に強力な体当たり攻撃", 8, 10, 10, 2, 1, 16, 0, 50, 0},
+    {"百なる一の剣", _S("↓↘→↗↑ため↓"), "インテリジェンス""\x83\x5C""ードが分身を召喚", 30, 20, 8, 130, 129, 0, 64, 20, 0},
+    {"", _S(""), "", 0, 0, 0, 0, 0, 0, 0, 0, 0},
+    {"", _S(""), "", 8, 10, 10, 2, 1, 0, 32768, 10, 0},
+    {"", _S(""), "", 0, 20, 20, 2, 1, 0, 0, 10, 0},
+    {"", _S(""), "", 0, 20, 20, 2, 129, 64, 64, 30, 0},
+    {"", _S(""), "", 0, 20, 20, 2, 1, 0, 128, 20, 0},
+    {"", _S(""), "", 5, 10, 10, 2, 1, 16, 0, 50, 0},
+    {"", _S(""), "", 0, 40, 10, 2, 129, 16, 64, 10, 0},
+    {"", _S(""), "", 0, 10, 4, 2, 1, 0, 0, 5, 0},
+    {"", _S(""), "", 0, 10, 10, 2, 1, 0, 32768, 10, 0},
+    {"", _S(""), "", 0, 2, 4, 2, 1, 0, 0, 5, 0},
+    {"", _S(""), "", 0, 2, 4, 2, 2, 0, 0, 5, 0},
+    {"", _S(""), "", 0, 20, 4, 2, 129, 0, 64, 5, 0},
+    {"", _S(""), "", 0, 20, 4, 2, 129, 0, 64, 10, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 64, 5, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 64, 10, 0},
+    {"", _S(""), "", 0, 4, 8, 2, 129, 0, 64, 5, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 32832, 30, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 8256, 30, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 16448, 30, 0},
+    {"", _S(""), "", 0, 20, 8, 2, 129, 0, 16448, 45, 0},
+};
+
+RelicDesc g_RelicDefs[] = {
+    {"\x83\x5C""ウルオブバット", "コウモリに変身する力を得る", 256, 256, 0},
+    {"ファイアオブバット", "コウモリ時に火炎を吐ける", 256, 257, 0},
+    {"エコーオブバット", "コウモリ時に暗闇を見通す超音波を出せる", 256, 258, 0},
+    {"クレイジーエコー", "超音波でも攻撃できるようになる", 256, 259, 0},
+    {"\x83\x5C""ウルオブウルフ", "狼に変身する力を得る", 257, 260, 0},
+    {"パワーオブウルフ", "狼時に速く走れる", 257, 261, 0},
+    {"スキルオブウルフ", "狼時に必殺技を出せる", 257, 262, 0},
+    {"ミスト", "霧に変身する力を得る", 258, 263, 0},
+    {"パワーオブミスト", "霧に変身できる時間が長くなる", 258, 264, 0},
+    {"ガスクラウド", "毒の霧になる", 258, 265, 0},
+    {"マテライズキューブ", "物品を破壊したときアイテムが出現", 259, 266, 0},
+    {"魂のオーブ", "敵に与えたダメージが見える", 260, 267, 0},
+    {"浮身のブーツ", "↓↑ジャンプでハイジャンプ（通常の５倍）", 261, 268, 0},
+    {"飛翔石", "二段ジャンプが可""\x94\x5C""　", 262, 269, 0},
+    {"ホーリーシンボル", "水でダメージを受けない", 263, 270, 0},
+    {"妖精の書", "敵の名前がわかる", 264, 271, 0},
+    {"アンロックジュエル", "青い光を放つ扉を開ける鍵", 266, 273, 0},
+    {"人魚の像", "地下水脈のとある場所に渡し守を呼べる", 272, 279, 0},
+    {"コウモリの素", "この場に使い魔「コウモリ」を召還する", 265, 272, 1},
+    {"ゴーストの素", "この場に使い魔「ゴースト」を召還する", 265, 272, 2},
+    {"妖精の素", "この場に使い魔「妖精」を召還する", 265, 272, 3},
+    {"子悪魔の素", "この場に使い魔「子悪魔」を召還する", 265, 272, 4},
+    {"剣魔の素", "使い魔「インテリジェンス""\x83\x5C""ード」を召還する", 265, 272, 5},
+    {"半妖精の素", "この場に使い魔「半妖精」を召還する", 265, 272, 6},
+    {"鼻悪魔の素", "この場に使い魔「鼻悪魔」を召還する", 265, 272, 7},
+    {"心臓", "魔導五器のひとつ　呪いダメージの無効化", 268, 275, 0},
+    {"爪", "魔導五器のひとつ　ＳＴＲ＋１０", 270, 277, 0},
+    {"あばら骨", "魔導五器のひとつ　ＣＯＮ＋１０", 267, 274, 0},
+    {"指輪", "魔導五器のひとつ　ＩＮＴ＋１０", 271, 278, 0},
+    {"目玉", "魔導五器のひとつ　ＬＣＫ＋１０", 269, 276, 0},
+    {"神速の靴", "→→（方向キー２回）でダッシュ可""\x94\x5C""　", 275, 281, 0},
+    {"縮地のブーツ", "宗次郎愛用のブーツ　目にも写らねぇ速さ", 269, 276, 0},
+};
+
+EnemyDef g_EnemyDefs[] = {
+    {_S(""), 32767, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xAD202000},
+    {_S(""), 32767, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 12, 0x0D202000},
+    {_S(""), 32767, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xCD202000},
+    {_S(""), 32767, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x8D202000},
+    {_S(""), 32767, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x05202000},
+    {_S(""), 32767, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("アックスアーマー"), 42, 10, 33, 6, 35, 0, 0, 0, 0, 9, 50, 136, 15, 2, 16, 8, 24, 0xAA00D410},
+    {_S(""), 0, 14, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D000010},
+    {_S(""), 200, 100, 2049, 200, 3, 0, 0, 0, 61280, 99, 500, 0, 0, 0, 0, 12, 24, 0x0800F410},
+    {_S("ヴィクトリーアーマー"), 61, 12, 33, 8, 35, 0, 0, 0, 0, 11, 80, 222, 218, 8, 16, 16, 27, 0xA800D410},
+    {_S(""), 32767, 23, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D20C000},
+    {_S("ブーメランスケルトン"), 18, 12, 33, 0, 35, 4096, 0, 0, 0, 5, 15, 264, 211, 4, 16, 5, 17, 0x28001410},
+    {_S(""), 32767, 18, 33, 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D200000},
+    {_S("ブラッディーゾンビ"), 24, 10, 65, 3, 51, 36864, 0, 0, 0, 4, 15, 146, 305, 16, 32, 4, 22, 0xA8001410},
+    {_S("フライングゾンビ"), 190, 37, 33, 10, 51, 36864, 0, 0, 0, 32, 50, 172, 205, 8, 16, 5, 8, 0x88001410},
+    {_S("フライングゾンビ"), 190, 37, 33, 10, 51, 36864, 0, 0, 0, 32, 50, 172, 205, 8, 16, 6, 11, 0x8A001010},
+    {_S("アンフィスバエナ"), 80, 6, 33, 0, 51, 36864, 1024, 0, 0, 7, 50, 168, 203, 6, 8, 24, 12, 0x28001410},
+    {_S(""), 32767, 6, 33, 0, 3, 36864, 1088, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0x0D202000},
+    {_S(""), 32767, 6, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D200000},
+    {_S(""), 0, 12, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10, 0x8D000010},
+    {_S("オウルアーマー"), 180, 17, 33, 2, 35, 0, 64, 0, 0, 15, 50, 388, 218, 8, 16, 9, 25, 0x0840D410},
+    {_S(""), 32767, 25, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ふくろう"), 26, 20, 33, 0, 51, 32768, 0, 0, 0, 10, 30, 0, 0, 0, 0, 12, 10, 0x08401810},
+    {_S("レッサーデーモン"), 400, 20, 33, 5, 51, 4096, 0, 2048, 0, 20, 100, 266, 237, 8, 8, 6, 20, 0x08401610},
+    {_S(""), 32767, 24, 35, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 36, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 44, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 3, 0x8D202000},
+    {_S("はんぎょじん"), 10, 13, 33, 0, 51, 49152, 1024, 0, 0, 2, 12, 129, 368, 8, 12, 5, 17, 0x8A001C10},
+    {_S(""), 0, 20, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 3, 0x8D000010},
+    {_S("はんぎょじん"), 10, 14, 33, 0, 51, 49152, 1024, 0, 0, 3, 12, 129, 368, 8, 16, 4, 21, 0x8A001C10},
+    {_S(""), 0, 24, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 3, 0x8D000010},
+    {_S("ゴーゴン"), 240, 35, 33, 40, 35, 0, 33376, 0, 0, 38, 555, 244, 277, 8, 20, 0, 0, 0x08603410},
+    {_S(""), 240, 55, 33, 2, 35, 9216, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x09402410},
+    {_S(""), 32767, 27, 514, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D202000},
+    {_S("グレートアーマー"), 84, 18, 33, 6, 35, 0, 64, 0, 0, 12, 100, 219, 151, 8, 16, 6, 24, 0x08003410},
+    {_S(""), 32767, 24, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202010},
+    {_S(""), 32767, 30, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ブラックパンサー"), 35, 45, 33, 0, 51, 32768, 0, 0, 0, 35, 600, 270, 200, 1, 16, 23, 6, 0x88003410},
+    {_S("クレイジーオクトパス"), 280, 45, 33, 2, 51, 49152, 1024, 0, 0, 38, 120, 186, 190, 10, 16, 26, 6, 0xAA001410},
+    {_S(""), 32767, 52, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D200200},
+    {_S("のみ男"), 11, 9, 33, 0, 51, 0, 0, 0, 0, 7, 17, 175, 155, 8, 16, 8, 8, 0xAA001810},
+    {_S("アーマードのみ男"), 18, 17, 33, 0, 51, 0, 0, 0, 0, 16, 40, 308, 290, 8, 10, 4, 11, 0xAA000410},
+    {_S(""), 32767, 25, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ホワイトドラゴン"), 260, 60, 33, 2, 3, 4096, 32768, 0, 0, 35, 150, 0, 0, 0, 0, 0, 0, 0x88001610},
+    {_S(""), 32767, 67, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0xCD200200},
+    {_S("こうきどうほねばしら"), 250, 20, 33, 4, 35, 36864, 0, 0, 0, 19, 40, 142, 131, 4, 16, 14, 8, 0x08403810},
+    {_S("こうきどうほねばしら"), 140, 20, 33, 2, 35, 36864, 0, 0, 0, 19, 30, 0, 131, 0, 4, 4, 18, 0x08403810},
+    {_S(""), 32767, 49, 16387, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 24, 24, 0x0D202000},
+    {_S("セムラー"), 17, 18, 33, 0, 51, 0, 0, 0, 0, 13, 50, 176, 197, 6, 16, 10, 14, 0x88001810},
+    {_S("マリオネット"), 20, 8, 33, 2, 35, 32768, 0, 0, 0, 6, 30, 341, 280, 16, 24, 3, 18, 0x28001410},
+    {_S("オルロック"), 666, 20, 33, 8, 35, 0, 2048, 0, 0, 25, 500, 0, 0, 0, 0, 5, 28, 0x08002010},
+    {_S(""), 13, 21, 33, 0, 51, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x09000210},
+    {_S(""), 15, 23, 33, 0, 35, 4096, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 5, 0x09000210},
+    {_S(""), 32767, 31, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 0, 0x0D202000},
+    {_S(""), 32767, 43, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 2, 0x0D202000},
+    {_S("オルロック"), 32767, 15, 33, 20, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x08602010},
+    {_S(""), 800, 15, 33, 4, 35, 4096, 0, 0, 0, 26, 2000, 0, 0, 0, 0, 0, 0, 0x09402010},
+    {_S(""), 32767, 23, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 28, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 0, 20, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D002000},
+    {_S(""), 32767, 35, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ワーパンサースケルトン"), 33, 20, 33, 4, 35, 4096, 0, 0, 0, 17, 50, 372, 278, 4, 24, 8, 16, 0xAA001410},
+    {_S(""), 0, 28, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0xCD000010},
+    {_S(""), 32767, 35, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("バット"), 1, 16, 33, 0, 55, 0, 0, 0, 0, 1, 10, 130, 289, 6, 24, 8, 8, 0x88001810},
+    {_S("スライム・ラージ"), 64, 32, 33, 10, 35, 32768, 0, 64, 0, 22, 100, 0, 0, 0, 0, 0, 0, 0x08403410},
+    {_S("スライム"), 32, 16, 33, 10, 35, 32768, 0, 64, 0, 16, 50, 0, 0, 0, 0, 0, 0, 0x08403410},
+    {_S("フライングスカル"), 15, 14, 33, 0, 35, 4096, 2048, 0, 0, 10, 30, 288, 334, 16, 16, 8, 8, 0xA8001410},
+    {_S("スケルトンモーニングスター"), 36, 14, 33, 2, 35, 4096, 0, 0, 0, 13, 50, 195, 259, 8, 16, 13, 22, 0xA8001410},
+    {_S(""), 32767, 24, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0x0D200000},
+    {_S("レッドスケルトン"), 9, 8, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 18, 0xAA000010},
+    {_S("アムドゥシアス"), 380, 17, 33, 8, 35, 0, 0, 0, 0, 23, 150, 314, 315, 3, 4, 14, 35, 0x0800F410},
+    {_S(""), 32767, 28, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 30, 16385, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 25, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("スケルトン"), 9, 2, 33, 0, 35, 4096, 0, 0, 0, 2, 10, 131, 282, 16, 24, 8, 18, 0xAA001810},
+    {_S(""), 0, 8, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD000010},
+    {_S("ディスクアーマー"), 450, 29, 33, 10, 35, 0, 0, 0, 0, 22, 140, 239, 263, 32, 16, 16, 32, 0x8A00D610},
+    {_S("ディスクアーマー"), 32767, 32, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0C20C200},
+    {_S("フレイムデーモン"), 320, 45, 33, 15, 51, 0, 0, 32768, 0, 38, 666, 247, 143, 4, 6, 12, 12, 0xAA003410},
+    {_S(""), 32767, 70, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D202000},
+    {_S("アルティバティラエ"), 18, 7, 33, 3, 35, 4096, 0, 0, 0, 3, 20, 0, 0, 0, 0, 6, 6, 0x08403810},
+    {_S(""), 32767, 5, 129, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 4, 0xCD202000},
+    {_S("スケルトンエイプ"), 10, 10, 33, 2, 35, 4096, 0, 0, 0, 10, 30, 209, 159, 8, 24, 8, 16, 0xAA001410},
+    {_S(""), 0, 20, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10, 0x0D000010},
+    {_S("しょくじんばな"), 60, 12, 33, 0, 35, 64, 1024, 0, 2048, 8, 60, 200, 133, 8, 16, 4, 4, 0x0C001610},
+    {_S(""), 0, 5, 514, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 0x8D000200},
+    {_S(""), 0, 6, 34, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 0x8D000200},
+    {_S("エクトプラズム"), 18, 6, 258, 0, 35, 36864, 64, 0, 0, 11, 70, 292, 275, 2, 32, 12, 12, 0xA8001410},
+    {_S("メディウサヘッド"), 1, 8, 33, 0, 39, 0, 512, 0, 0, 4, 4, 141, 286, 1, 16, 8, 8, 0xC8001810},
+    {_S("ほねばしら"), 16, 0, 33, 8, 34, 4096, 0, 0, 0, 1, 2, 0, 0, 0, 0, 16, 12, 0x28001810},
+    {_S(""), 0, 10, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0xCD000010},
+    {_S(""), 32767, 16, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 8, 0x0D200000},
+    {_S("アーマーナイト"), 20, 12, 33, 4, 35, 4096, 0, 0, 0, 10, 70, 308, 212, 12, 16, 8, 20, 0x9A00D410},
+    {_S(""), 32767, 14, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D20C000},
+    {_S(""), 32767, 16, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D20C000},
+    {_S(""), 32767, 25, 65, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 4, 0xCD600000},
+    {_S("ヘビーアーマー"), 90, 18, 33, 6, 35, 0, 0, 0, 0, 8, 90, 201, 265, 8, 16, 0, 0, 0x08403410},
+    {_S(""), 32767, 35, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D602000},
+    {_S("フローズンシェイド"), 16, 16, 8193, 0, 35, 32768, 0, 1024, 8192, 12, 40, 171, 316, 6, 12, 4, 20, 0xA8001410},
+    {_S(""), 32767, 23, 8193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D200000},
+    {_S(""), 4, 10, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 4, 0x0D202000},
+    {_S("スケルトン・ガンマン"), 24, 12, 33, 5, 35, 4096, 0, 0, 0, 8, 20, 389, 153, 1, 8, 8, 15, 0x08001A10},
+    {_S(""), 32767, 45, 33, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0x0D202000},
+    {_S("ドードー"), 2, 1, 33, 0, 35, 0, 0, 0, 0, 49, 111, 273, 383, 2, 8, 8, 8, 0x88003410},
+    {_S("スケルトンソルジャー"), 18, 5, 33, 2, 35, 0, 0, 0, 0, 3, 15, 154, 147, 16, 32, 8, 18, 0xA8001410},
+    {_S("トード"), 10, 14, 33, 4, 51, 49152, 1024, 0, 0, 11, 20, 174, 267, 8, 8, 8, 10, 0xA8001810},
+    {_S("フロッグ"), 2, 13, 33, 1, 51, 49152, 1024, 0, 0, 12, 20, 174, 215, 8, 24, 8, 7, 0xA8001810},
+    {_S("アーチャースケルトン"), 10, 12, 33, 4, 35, 0, 0, 0, 0, 10, 50, 153, 7, 16, 32, 8, 10, 0xA8001410},
+    {_S(""), 1, 20, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 2, 0x8D000010},
+    {_S("ゾンビ"), 1, 14, 33, 0, 35, 36864, 0, 0, 0, 1, 5, 305, 5, 24, 16, 10, 16, 0x88001C10},
+    {_S("グレイブ・キーパー"), 123, 20, 33, 0, 51, 57344, 0, 0, 0, 18, 60, 187, 189, 6, 8, 10, 20, 0x88001610},
+    {_S(""), 32767, 25, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("グレイブ"), 5, 40, 257, 100, 35, 32, 64, 512, 0, 25, 88, 186, 230, 6, 16, 10, 20, 0xA8001410},
+    {_S("ブルークロウ"), 15, 10, 33, 0, 51, 0, 16384, 0, 0, 13, 50, 191, 368, 6, 24, 0, 0, 0xA8001810},
+    {_S("ブラッククロウ"), 15, 10, 33, 0, 51, 0, 16384, 0, 0, 13, 50, 192, 369, 6, 24, 0, 0, 0xA8001810},
+    {_S("スローマスター"), 20, 40, 33, 4, 35, 4096, 0, 0, 0, 33, 150, 208, 205, 4, 8, 8, 18, 0xA8001410},
+    {_S(""), 0, 30, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0xCD000210},
+    {_S("スケルトンスピア"), 30, 12, 33, 2, 35, 4096, 0, 0, 0, 14, 40, 176, 212, 6, 8, 5, 19, 0xA8001410},
+    {_S(""), 32767, 25, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("サッカーボーイ"), 10, 20, 33, 2, 35, 0, 0, 0, 0, 29, 300, 142, 131, 16, 32, 4, 17, 0xA8001410},
+    {_S("ボール"), 2, 0, 33, 0, 3, 0, 0, 0, 0, 0, 0, 370, 265, 16, 16, 2, 2, 0x08000410},
+    {_S("スケルトンブレイズ"), 65, 20, 33, 4, 35, 0, 0, 0, 0, 17, 80, 206, 156, 4, 8, 4, 17, 0xA8001410},
+    {_S(""), 32767, 25, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ブレイズ・ノーヴィス"), 16, 15, 33, 4, 35, 0, 0, 0, 0, 8, 20, 214, 7, 16, 32, 5, 18, 0xA8001410},
+    {_S(""), 32767, 20, 65, 0, 1, 0, 0, 0, 0, 0, 0, 221, 214, 4, 8, 0, 0, 0x0D202000},
+    {_S("スケルトン・ビーマー"), 20, 20, 33, 4, 35, 0, 0, 0, 0, 35, 444, 246, 131, 6, 48, 5, 18, 0x88003410},
+    {_S(""), 32767, 150, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("フライングスケルトン"), 15, 12, 33, 0, 35, 0, 0, 0, 0, 10, 30, 212, 135, 4, 10, 4, 20, 0x88001810},
+    {_S("ルーラーソード"), 90, 25, 33, 4, 35, 0, 0, 0, 0, 13, 80, 226, 221, 4, 16, 1, 29, 0x08401610},
+    {_S("ポルターガイスト"), 30, 20, 33, 2, 3, 0, 0, 0, 0, 8, 10, 0, 210, 0, 8, 0, 0, 0x08000610},
+    {_S("ビフロン"), 99, 18, 33, 0, 51, 0, 0, 0, 0, 20, 50, 241, 196, 8, 12, 8, 16, 0xA8001410},
+    {_S(""), 0, 29, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0xCD000000},
+    {_S("デッドトルーパー"), 161, 16, 33, 0, 35, 4096, 0, 0, 0, 19, 100, 228, 225, 8, 16, 0, 0, 0x08003410},
+    {_S(""), 32766, 21, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 38, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ルーラーソードLv2"), 100, 25, 33, 6, 35, 0, 0, 0, 0, 15, 100, 226, 221, 4, 16, 1, 29, 0x08401610},
+    {_S("パペットソード"), 20, 15, 65, 4, 3, 0, 0, 0, 0, 10, 10, 0, 155, 0, 16, 0, 0, 0x08000610},
+    {_S("ルーラーソードLv3"), 540, 55, 33, 4, 35, 0, 0, 0, 0, 36, 400, 254, 251, 2, 16, 1, 29, 0x08401610},
+    {_S("スピア"), 10, 65, 65, 2, 3, 0, 0, 0, 0, 20, 300, 0, 212, 0, 32, 4, 4, 0x08000610},
+    {_S("シールド"), 120, 20, 33, 40, 3, 0, 0, 0, 0, 20, 200, 0, 134, 0, 32, 4, 16, 0x0800C610},
+    {_S("ケツアルクアトル"), 200, 47, 33, 4, 35, 0, 0, 0, 0, 35, 150, 375, 152, 4, 16, 7, 13, 0x08003410},
+    {_S("ケツアルクアトル"), 32767, 47, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 6, 0x0C202000},
+    {_S("ケツアルクアトルライダー"), 100, 31, 33, 0, 51, 0, 0, 0, 0, 35, 150, 0, 0, 0, 0, 5, 10, 0x08002010},
+    {_S("ホワイトドラゴンライダー"), 120, 41, 33, 4, 35, 0, 0, 0, 0, 35, 150, 0, 0, 0, 0, 7, 13, 0x08001610},
+    {_S("ホワイトドラゴンライダー"), 32767, 41, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 6, 0x0C200200},
+    {_S("デュラハンスケルトン"), 32, 7, 33, 3, 35, 4096, 0, 0, 0, 12, 50, 151, 306, 8, 32, 0, 0, 0x28001410},
+    {_S(""), 32767, 15, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("フェンリル"), 200, 41, 33, 0, 51, 8192, 0, 32768, 0, 36, 160, 152, 370, 8, 20, 0, 0, 0x08001410},
+    {_S(""), 32767, 48, 65, 0, 51, 8192, 0, 32768, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x09002000},
+    {_S(""), 32767, 60, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D200000},
+    {_S("フェンリルライダー"), 120, 36, 33, 3, 35, 8192, 64, 32768, 0, 37, 160, 0, 0, 0, 0, 0, 0, 0x08401010},
+    {_S(""), 32767, 55, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("チュパカパブラ"), 88, 35, 33, 5, 51, 32768, 1024, 0, 0, 38, 333, 393, 201, 16, 32, 0, 0, 0x28003410},
+    {_S(""), 32767, 50, 65, 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D202000},
+    {_S(""), 32767, 69, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D202000},
+    {_S("ゴースト"), 11, 5, 33, 0, 39, 4096, 2048, 0, 0, 6, 10, 274, 7, 8, 16, 7, 7, 0xC8001C10},
+    {_S("ウネ"), 12, 10, 33, 0, 35, 32768, 32, 0, 0, 5, 20, 161, 160, 8, 16, 0, 0, 0x88001810},
+    {_S("じんめんそう"), 18, 20, 33, 0, 35, 32768, 32, 0, 0, 13, 100, 289, 274, 4, 8, 0, 0, 0x88001410},
+    {_S("じんめんそう"), 20, 20, 33, 0, 35, 32768, 0, 0, 0, 13, 100, 289, 274, 8, 16, 5, 8, 0x08001410},
+    {_S(""), 0, 28, 129, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0xCD000000},
+    {_S("アルラ・ウネ"), 100, 20, 33, 0, 35, 32768, 32, 0, 0, 25, 150, 272, 348, 8, 32, 12, 7, 0x08003410},
+    {_S("アルラ・ウネ"), 500, 28, 33, 0, 35, 32768, 0, 0, 0, 25, 150, 272, 348, 2, 8, 10, 34, 0x08003410},
+    {_S(""), 32767, 36, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 161, 0, 16, 0, 0, 0x0D202010},
+    {_S(""), 1, 40, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 157, 0, 16, 3, 3, 0x8D000210},
+    {_S("ボンバーアーマー"), 46, 37, 33, 20, 35, 0, 0, 0, 0, 30, 140, 268, 209, 8, 16, 0, 0, 0xA8001410},
+    {_S(""), 25, 70, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0x8D000010},
+    {_S("ロックアーマー"), 160, 50, 33, 20, 35, 0, 0, 0, 0, 36, 250, 312, 227, 8, 16, 0, 0, 0xA8001410},
+    {_S(""), 10, 100, 33, 22, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D000010},
+    {_S("ドラキュラ"), 600, 4, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 5, 0x0D202010},
+    {_S(""), 0, 7, 32769, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0xCD002010},
+    {_S(""), 32767, 10, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0xCD002000},
+    {_S("グレーターデーモン"), 1200, 8, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x0D202010},
+    {_S(""), 32767, 12, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD002000},
+    {_S(""), 8, 10, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 13, 0x0D202010},
+    {_S("アンフアゥグリア"), 32, 20, 33, 0, 51, 32768, 0, 0, 0, 2, 10, 0, 0, 0, 0, 0, 0, 0xA8001410},
+    {_S(""), 32767, 28, 65, 0, 51, 32768, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x09002000},
+    {_S("ゴースト"), 1, 10, 33, 0, 39, 4096, 0, 0, 0, 1, 8, 162, 163, 12, 48, 7, 7, 0xC8001810},
+    {_S("スケルトン・リブ"), 12, 5, 33, 0, 35, 4096, 0, 0, 0, 4, 10, 134, 133, 4, 24, 8, 18, 0xA8001410},
+    {_S(""), 0, 8, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD000010},
+    {_S("スラントガーダー"), 48, 13, 33, 3, 35, 0, 0, 0, 0, 12, 30, 223, 218, 4, 16, 7, 22, 0xA8001410},
+    {_S(""), 32767, 16, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D200000},
+    {_S("ちょんちょん"), 4, 55, 2049, 0, 51, 0, 0, 0, 0, 33, 128, 382, 279, 16, 48, 10, 16, 0x08001410},
+    {_S("ほねばしら"), 64, 12, 33, 2, 35, 4096, 0, 0, 0, 10, 30, 332, 274, 12, 13, 12, 12, 0x28001410},
+    {_S(""), 32767, 23, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 21, 24, 0x0D202000},
+    {_S("じごくのてっきゅう"), 500, 59, 33, 28, 3, 0, 0, 0, 0, 40, 100, 329, 265, 1, 255, 13, 13, 0x08003410},
+    {_S("ハンマーシェイカー"), 250, 20, 33, 0, 35, 0, 0, 0, 0, 21, 110, 311, 277, 4, 32, 4, 6, 0x08001610},
+    {_S(""), 32767, 40, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("グルカ・マスター"), 420, 24, 33, 0, 35, 0, 0, 0, 0, 21, 120, 311, 148, 4, 8, 4, 6, 0x08001610},
+    {_S(""), 32767, 30, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D200200},
+    {_S("レッドスマッシャー"), 380, 24, 33, 0, 35, 0, 0, 0, 0, 21, 100, 311, 224, 4, 16, 4, 6, 0x08001610},
+    {_S(""), 32767, 25, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 28, 35, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S("ターンネートテーブル"), 20, 4, 33, 0, 35, 0, 0, 0, 0, 5, 20, 178, 185, 8, 8, 7, 13, 0x28001410},
+    {_S(""), 32767, 5, 33, 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 0x0D200000},
+    {_S("アマラリックスナイパー"), 50, 40, 33, 10, 35, 2048, 4096, 0, 0, 36, 200, 324, 153, 2, 32, 20, 23, 0x88003410},
+    {_S(""), 0, 65, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0xCD002010},
+    {_S(""), 0, 55, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0xCD002010},
+    {_S("ガラモス"), 12000, 50, 33, 56, 35, 0, 0, 0, 0, 50, 9999, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S("ガラモス"), 32767, 50, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 92, 33, 0, 9, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x0D202000},
+    {_S(""), 32767, 98, 16386, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD202000},
+    {_S(""), 32767, 85, 36, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D200000},
+    {_S("ミノタウロス"), 300, 20, 33, 3, 51, 0, 0, 0, 0, 18, 400, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 35, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 13, 13, 0x0D202000},
+    {_S(""), 32767, 15, 129, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D200200},
+    {_S("ウェアウルフ"), 260, 20, 33, 1, 51, 0, 0, 0, 0, 18, 300, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 28, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 30, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D200200},
+    {_S(""), 32767, 40, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D200200},
+    {_S(""), 0, 18, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD200000},
+    {_S("ビッグ・がいこつ"), 100, 24, 33, 0, 35, 4096, 0, 0, 0, 16, 50, 379, 385, 8, 16, 8, 8, 0x08001410},
+    {_S(""), 32767, 30, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D202000},
+    {_S(""), 32767, 20, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0xCD000010},
+    {_S("マッドマン"), 15, 20, 33, 0, 35, 0, 0, 0, 0, 15, 50, 0, 0, 0, 0, 8, 16, 0x88001410},
+    {_S(""), 32767, 30, 67, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("スケルトンジャンパー"), 30, 56, 33, 0, 35, 4096, 0, 0, 0, 37, 160, 340, 207, 16, 24, 5, 25, 0x88001410},
+    {_S("フローズンハーフ"), 118, 35, 8193, 2, 35, 32768, 1024, 0, 8192, 45, 600, 344, 384, 6, 16, 5, 22, 0x08001410},
+    {_S(""), 32767, 60, 8193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 0x0D200000},
+    {_S(""), 32767, 75, 8193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x8D200000},
+    {_S(""), 32767, 65, 8193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 32, 32, 0x0D200000},
+    {_S("アリオルムナス"), 180, 20, 33, 0, 35, 0, 0, 0, 0, 20, 80, 167, 342, 5, 8, 5, 22, 0x08001410},
+    {_S(""), 32767, 25, 257, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D000010},
+    {_S(""), 32767, 35, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D000210},
+    {_S("ブーロー"), 330, 80, 65, 22, 35, 0, 2080, 0, 0, 45, 700, 252, 392, 4, 12, 0, 0, 0x08001610},
+    {_S("ユコバック"), 100, 20, 33, 0, 35, 8192, 32768, 0, 0, 18, 60, 314, 283, 2, 32, 10, 13, 0x08001410},
+    {_S(""), 0, 24, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 6, 0x0D000010},
+    {_S("ネメシス"), 88, 18, 65, 0, 35, 0, 4128, 0, 0, 15, 70, 166, 150, 6, 12, 0, 0, 0x28001410},
+    {_S("アラストール"), 120, 30, 33, 3, 35, 0, 32, 0, 0, 15, 100, 271, 245, 2, 4, 6, 6, 0x08001410},
+    {_S("サウジーネ"), 210, 50, 33, 2, 35, 0, 0, 0, 0, 45, 450, 292, 351, 6, 8, 8, 25, 0x08001410},
+    {_S(""), 0, 60, 33, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D000010},
+    {_S(""), 2, 65, 33, 25, 11, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 8, 0x8D000010},
+    {_S(""), 25, 60, 2049, 0, 39, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 5, 0x8D000010},
+    {_S("デビル"), 200, 24, 33, 2, 35, 4096, 2048, 0, 0, 9, 100, 204, 203, 3, 5, 19, 38, 0x08001410},
+    {_S(""), 0, 38, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D000010},
+    {_S(""), 32767, 48, 8193, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 13, 44, 0x0D000010},
+    {_S("エビル"), 450, 52, 33, 20, 35, 4096, 0, 0, 2048, 39, 666, 320, 138, 4, 8, 19, 38, 0x08003410},
+    {_S(""), 32767, 85, 129, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 25, 14, 0x0D202000},
+    {_S(""), 32767, 65, 2049, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 22, 22, 0x0D202000},
+    {_S("ハルピュイアイ"), 26, 18, 33, 0, 35, 0, 16384, 0, 0, 18, 70, 276, 158, 2, 16, 8, 20, 0x08001410},
+    {_S(""), 32767, 24, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0x0D202000},
+    {_S(""), 0, 26, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D000010},
+    {_S(""), 0, 20, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x8D000010},
+    {_S("ベリガン"), 200, 6, 33, 0, 35, 0, 0, 0, 0, 6, 200, 0, 0, 0, 0, 0, 0, 0x08001210},
+    {_S(""), 32767, 9, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 0, 10, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D000010},
+    {_S("アックスアーマー"), 32, 6, 33, 3, 35, 32, 0, 0, 0, 4, 10, 307, 15, 4, 16, 8, 24, 0xAA001410},
+    {_S("じゅほうてん"), 26, 7, 33, 0, 35, 32768, 32, 0, 0, 10, 30, 203, 9, 2, 8, 12, 12, 0x08001410},
+    {_S(""), 32767, 14, 35, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 0x0D202000},
+    {_S("まほうてん"), 50, 8, 33, 0, 35, 32768, 32, 0, 0, 8, 18, 0, 0, 0, 0, 12, 12, 0x08001410},
+    {_S(""), 32767, 16, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 0x0D202000},
+    {_S("ぶぐじてん"), 22, 12, 33, 2, 35, 32768, 32, 0, 0, 12, 35, 219, 10, 4, 4, 12, 12, 0x08001410},
+    {_S(""), 32767, 20, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 0x0D202000},
+    {_S("ドッペルゲンガーLV10"), 120, 7, 33, 5, 51, 0, 0, 0, 0, 10, 500, 0, 0, 0, 0, 0, 0, 0x08000010},
+    {_S("ギャイボン"), 200, 7, 33, 0, 35, 0, 0, 0, 0, 6, 200, 0, 0, 0, 0, 0, 0, 0x08001210},
+    {_S(""), 0, 8, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D000010},
+    {_S(""), 0, 12, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10, 0x8D000010},
+    {_S(""), 4096, 6, 4097, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D000010},
+    {_S(""), 32767, 15, 37, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 9, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 5, 513, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S("だいぼうれい"), 80, 25, 33, 4, 35, 4096, 2048, 0, 0, 12, 50, 142, 217, 2, 16, 24, 32, 0x08001410},
+    {_S("ライオン"), 150, 37, 33, 25, 35, 0, 0, 0, 0, 27, 1000, 250, 385, 6, 16, 18, 29, 0x28001410},
+    {_S(""), 32767, 77, 67, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ティンマン"), 48, 32, 33, 36, 35, 0, 64, 0, 0, 28, 1000, 325, 179, 4, 16, 7, 24, 0x08001410},
+    {_S(""), 32767, 55, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D202000},
+    {_S(""), 0, 33, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9, 3, 0x8D000210},
+    {_S("ミイラ男"), 1200, 40, 33, 0, 35, 36864, 0, 896, 2048, 40, 2500, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 66, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 1, 129, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x8D200200},
+    {_S(""), 0, 71, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S("フライングアーマー"), 65, 20, 33, 3, 35, 0, 0, 0, 0, 20, 80, 249, 231, 1, 16, 7, 26, 0x08401810},
+    {_S(""), 32767, 30, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D200000},
+    {_S("おおコウモリ"), 600, 35, 33, 0, 35, 0, 0, 0, 0, 35, 1200, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 0, 65, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S(""), 0, 55, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S(""), 0, 45, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S("フィッシュヘッド"), 70, 15, 33, 2, 35, 4096, 32768, 0, 0, 20, 50, 243, 285, 8, 32, 15, 12, 0x08001410},
+    {_S(""), 0, 32, 32769, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 4, 0x8D000010},
+    {_S(""), 32767, 37, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("マルファス"), 500, 20, 33, 10, 35, 0, 2048, 0, 0, 22, 1000, 387, 288, 2, 24, 0, 0, 0x0A003410},
+    {_S(""), 0, 24, 33, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D002010},
+    {_S(""), 32767, 28, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x0D202000},
+    {_S(""), 20, 32, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 6, 0x8D002010},
+    {_S("インプ"), 43, 10, 33, 0, 35, 0, 0, 0, 0, 41, 66, 391, 279, 4, 32, 8, 8, 0x08401410},
+    {_S("バルーン"), 3, 55, 33, 0, 35, 0, 0, 0, 0, 29, 88, 0, 0, 0, 0, 7, 7, 0x08001410},
+    {_S(""), 0, 2, 129, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 3, 0x8D000010},
+    {_S("スキュラ"), 200, 16, 33, 0, 35, 16384, 0, 0, 1024, 14, 500, 0, 0, 0, 0, 0, 0, 0x08003010},
+    {_S(""), 32767, 18, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0C002010},
+    {_S(""), 32767, 20, 1025, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D200200},
+    {_S(""), 4, 8, 1025, 0, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D000210},
+    {_S(""), 32767, 20, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 0, 18, 1025, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x8D000210},
+    {_S(""), 8, 12, 33, 0, 11, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D000210},
+    {_S("スキュラテイル"), 130, 14, 33, 3, 35, 16384, 0, 1024, 0, 10, 100, 0, 0, 0, 0, 8, 8, 0x08401210},
+    {_S("レギオン"), 400, 30, 33, 10, 35, 4096, 2048, 0, 0, 28, 3000, 0, 0, 0, 0, 20, 20, 0x08403010},
+    {_S("レギオン"), 400, 28, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 10, 20, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x8D000210},
+    {_S(""), 32767, 55, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D202000},
+    {_S(""), 32767, 6, 65, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D000010},
+    {_S("ヒポグリフ"), 800, 18, 33, 0, 35, 0, 16384, 0, 0, 16, 800, 0, 0, 0, 0, 20, 6, 0x08403010},
+    {_S(""), 20, 10, 33, 0, 35, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x0D000210},
+    {_S(""), 32767, 28, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 0x0D202000},
+    {_S("メディウサヘッド"), 12, 12, 33, 0, 51, 0, 0, 512, 0, 7, 20, 141, 286, 1, 10, 4, 4, 0x88001810},
+    {_S("メディウサヘッド"), 12, 7, 513, 0, 51, 0, 0, 0, 512, 8, 30, 141, 286, 1, 10, 4, 4, 0x88001810},
+    {_S("アーチャー"), 300, 40, 33, 8, 35, 0, 0, 0, 0, 34, 140, 293, 272, 1, 4, 7, 20, 0x08403410},
+    {_S(""), 1, 55, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D002010},
+    {_S("リヒター・ベルモンド"), 400, 45, 33, 10, 51, 2048, 0, 4608, 0, 48, 0, 0, 0, 0, 0, 0, 0, 0x08002010},
+    {_S(""), 32767, 35, 4097, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 60, 36865, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 55, 37, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 35, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 4097, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 30, 4098, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 36866, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 40, 4161, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 35, 36866, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 40, 4163, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 30, 4098, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 40, 4099, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S("スケアクロウ"), 120, 40, 33, 6, 35, 0, 0, 0, 0, 27, 1000, 271, 212, 1, 24, 0, 0, 0x08403410},
+    {_S("キュウ"), 50, 40, 2049, 0, 35, 36864, 0, 0, 0, 28, 1000, 294, 188, 1, 8, 10, 10, 0x08003410},
+    {_S("ベルゼブブ"), 2000, 60, 2049, 13, 35, 4096, 0, 0, 2048, 44, 4444, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 10, 66, 2049, 0, 35, 4096, 2048, 0, 0, 0, 0, 0, 0, 0, 0, 20, 15, 0x0D000210},
+    {_S(""), 0, 5, 129, 0, 47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D000210},
+    {_S(""), 10, 0, 33, 5000, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x0D000010},
+    {_S("ラルフ・フェイク"), 1200, 38, 33, 8, 35, 0, 0, 0, 0, 41, 1000, 0, 0, 0, 0, 0, 0, 0x08003010},
+    {_S(""), 32767, 43, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 38, 2049, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x0D200200},
+    {_S(""), 32767, 58, 2049, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x0D200200},
+    {_S(""), 32767, 37, 65, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 2, 0x0D200200},
+    {_S(""), 32767, 18, 4097, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("グラント・フェイク"), 800, 38, 33, 4, 35, 0, 0, 0, 0, 41, 1200, 0, 0, 0, 0, 0, 0, 0x08003010},
+    {_S(""), 32767, 58, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 33, 65, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 7, 0x0D200200},
+    {_S("サイファ・フェイク"), 1000, 9, 33, 2, 35, 0, 0, 0, 0, 42, 1500, 0, 0, 0, 0, 0, 0, 0x08003010},
+    {_S(""), 32767, 60, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 35, 8193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 45, 4097, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D200200},
+    {_S(""), 180, 40, 257, 2, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 20, 0x0D200200},
+    {_S("サキュバス"), 666, 25, 33, 0, 35, 0, 2048, 0, 0, 25, 2000, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 0, 17, 257, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 3, 0x8D200200},
+    {_S(""), 0, 25, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x09402010},
+    {_S(""), 32767, 25, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 5, 0x0D200200},
+    {_S(""), 32767, 30, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D202000},
+    {_S(""), 0, 35, 2049, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x00000000},
+    {_S(""), 32767, 70, 4098, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S(""), 32767, 40, 4098, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D000010},
+    {_S("キラーフィッシュ"), 120, 50, 33, 4, 35, 16384, 33792, 0, 0, 25, 100, 190, 369, 2, 32, 0, 0, 0xA8001410},
+    {_S("シャフト"), 1300, 40, 33, 30, 35, 0, 0, 128, 0, 88, 0, 0, 0, 0, 0, 26, 26, 0x08403010},
+    {_S(""), 32767, 30, 33, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x0920F010},
+    {_S(""), 32767, 42, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 10, 0x4D202000},
+    {_S(""), 32767, 43, 16385, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 0x8D202000},
+    {_S(""), 32767, 64, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D202000},
+    {_S("デス"), 888, 35, 33, 17, 35, 4096, 0, 0, 2048, 66, 4444, 0, 0, 0, 0, 17, 37, 0x08403010},
+    {_S(""), 0, 55, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 0x8D000200},
+    {_S(""), 32767, 60, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 15, 15, 0x09203010},
+    {_S(""), 32767, 70, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 17, 0x09203010},
+    {_S(""), 32767, 80, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x09203010},
+    {_S("デス"), 1200, 40, 33, 14, 35, 4096, 0, 0, 2048, 66, 1000, 0, 0, 0, 0, 6, 6, 0x08403010},
+    {_S(""), 32767, 75, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 0x09203010},
+    {_S("ケルベロス"), 800, 20, 33, 0, 35, 8192, 0, 0, 32768, 24, 1500, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 35, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 4, 0x8D200200},
+    {_S(""), 32767, 40, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0x8D200200},
+    {_S("メディウサ"), 1100, 35, 33, 8, 35, 0, 32, 0, 512, 40, 2500, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 45, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D20F010},
+    {_S(""), 32767, 13, 514, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 0x09203010},
+    {_S(""), 32767, 50, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x09203010},
+    {_S("フランケン"), 1100, 30, 33, 15, 35, 0, 0, 16384, 0, 40, 2500, 0, 0, 0, 0, 0, 0, 0x08403010},
+    {_S(""), 32767, 55, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x09203010},
+    {_S("ドッペルゲンガーLV40"), 777, 35, 33, 16, 51, 0, 0, 0, 0, 40, 2001, 0, 0, 0, 0, 0, 0, 0x08000010},
+    {_S(""), 4096, 12, 4097, 0, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x0D000010},
+    {_S(""), 32767, 55, 37, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 35, 193, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 5, 513, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 35, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S(""), 32767, 40, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 12, 12, 0x0D000010},
+    {_S("ドラキュラ"), 10000, 39, 33, 18, 3, 0, 49152, 128, 2048, 98, 0, 0, 0, 0, 0, 58, 22, 0x08403010},
+    {_S(""), 32767, 70, 33, 20, 3, 0, 49152, 128, 2048, 0, 0, 0, 0, 0, 0, 12, 24, 0x09402010},
+    {_S(""), 32767, 85, 33, 30, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 12, 0x09402010},
+    {_S(""), 32767, 140, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 16, 0x09402010},
+    {_S(""), 32767, 95, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 14, 14, 0x09402010},
+    {_S("スピンヘッド"), 32767, 60, 33, 0, 3, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 10, 10, 0x08402010},
+    {_S(""), 32767, 12, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D602000},
+    {_S("ミノタウロス"), 320, 45, 33, 4, 51, 0, 0, 0, 0, 35, 250, 326, 196, 4, 16, 0, 0, 0x08403410},
+    {_S(""), 32767, 65, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 13, 13, 0x0D202000},
+    {_S(""), 32767, 5, 129, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D200200},
+    {_S("ウェアウルフ"), 280, 40, 33, 1, 51, 0, 0, 0, 0, 34, 200, 295, 232, 1, 8, 0, 0, 0x0A403410},
+    {_S(""), 32767, 45, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 72, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 4, 0x8D200200},
+    {_S("ブルー・アルラ・ウネ"), 100, 35, 33, 0, 35, 32768, 32, 0, 0, 40, 1000, 272, 233, 4, 16, 12, 7, 0x08003410},
+    {_S("ブルー・アルラ・ウネ"), 550, 45, 33, 1, 35, 32768, 0, 0, 0, 40, 1000, 272, 233, 4, 16, 10, 34, 0x08003410},
+    {_S(""), 32767, 50, 65, 0, 3, 0, 0, 0, 0, 0, 0, 0, 161, 0, 16, 0, 0, 0x0D202010},
+    {_S(""), 1, 70, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 157, 0, 16, 3, 3, 0x8D002010},
+    {_S("ファイナル・ガード"), 500, 50, 33, 34, 35, 0, 64, 57344, 2048, 60, 1500, 328, 257, 2, 1, 6, 24, 0x08403410},
+    {_S(""), 32767, 120, 65, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202010},
+    {_S(""), 32767, 90, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 0, 8, 65, 0, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D000010},
+    {_S("ガーゴイル"), 50, 20, 2049, 0, 35, 0, 0, 0, 0, 15, 50, 368, 5, 24, 16, 10, 16, 0x08003C10},
+    {_S("カッパーアーマー"), 230, 45, 33, 10, 35, 0, 64, 0, 0, 30, 300, 0, 0, 0, 0, 6, 24, 0x0840F010},
+    {_S("イフリート"), 330, 80, 32769, 15, 35, 8192, 0, 32768, 0, 46, 700, 247, 314, 3, 5, 6, 24, 0x08403410},
+    {_S(""), 32767, 80, 32769, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0xCD202000},
+    {_S("スケルトン・リーダー"), 444, 34, 33, 20, 35, 4096, 0, 0, 0, 36, 250, 0, 0, 0, 0, 8, 8, 0x08403010},
+    {_S("ガーデナー"), 130, 20, 33, 10, 51, 32768, 0, 0, 0, 20, 120, 163, 157, 12, 24, 10, 16, 0x08003410},
+    {_S("じんめんじゅ"), 250, 30, 33, 11, 2, 4096, 0, 0, 2048, 21, 125, 274, 158, 3, 16, 6, 6, 0x08003410},
+    {_S(""), 1, 25, 33, 1, 15, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D000010},
+    {_S("マリア・ラーネッド"), 400, 40, 33, 15, 51, 2048, 0, 4608, 0, 48, 0, 0, 0, 0, 0, 0, 0, 0x08402010},
+    {_S(""), 32767, 25, 32769, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 25, 16385, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 25, 8193, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S(""), 32767, 25, 4097, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x0D202000},
+    {_S("ウォーターリーパー"), 10, 14, 33, 0, 51, 49152, 1024, 0, 0, 3, 12, 0, 0, 0, 0, 4, 21, 0x8A001010},
+    {_S("ヒトトリソウ"), 128, 20, 33, 5, 35, 32768, 0, 0, 0, 20, 100, 301, 159, 8, 24, 6, 20, 0x08003410},
+    {_S("ガーゴイル"), 80, 30, 2049, 0, 35, 0, 0, 0, 0, 25, 150, 286, 7, 10, 32, 10, 16, 0x08003410},
+    {_S("ガーゴイル"), 110, 40, 2049, 10, 35, 0, 0, 0, 0, 35, 250, 303, 277, 3, 20, 10, 16, 0x08003410},
+    {_S("スペクター"), 20, 10, 2049, 0, 35, 4096, 57472, 0, 2400, 25, 120, 292, 274, 6, 13, 10, 16, 0x08403410},
+    {_S("ウィル・オ・ウィスプ"), 4, 8, 2049, 0, 35, 0, 0, 0, 0, 23, 90, 292, 289, 6, 48, 10, 16, 0x08003410},
+    {_S(""), 0, 65, 4097, 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x8D000010},
+    {_S("スケルトン・ブリーダー"), 12, 5, 33, 0, 35, 4096, 0, 0, 0, 4, 10, 0, 0, 0, 0, 8, 18, 0x08003010},
+    {_S("スケルトン・ビースト"), 6, 3, 33, 0, 35, 4096, 0, 0, 0, 3, 8, 0, 0, 0, 0, 8, 18, 0x08003010},
+    {_S(""), 32767, 10, 33, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 8, 0x0D000010},
+    {_S("ワイト"), 280, 20, 33, 10, 35, 4096, 0, 0, 0, 34, 210, 142, 138, 2, 8, 10, 16, 0x08403410},
+    {_S("レイス"), 300, 33, 33, 5, 35, 4096, 0, 0, 0, 32, 220, 271, 203, 1, 8, 10, 16, 0x08403410},
+    {_S("スケルトン・ガーディアン"), 320, 35, 33, 0, 35, 4096, 0, 0, 0, 37, 500, 0, 0, 0, 0, 8, 18, 0x08403010},
+};
+
+s32 g_ExpNext[] = {
+    0,      0,      100,    250,    450,    700,    1000,   1350,   1750,
+    2200,   2700,   3250,   3850,   4500,   5200,   5950,   6750,   7600,
+    8500,   9450,   10450,  11700,  13200,  15100,  17500,  20400,  23700,
+    27200,  30900,  35000,  39500,  44500,  50000,  56000,  61500,  68500,
+    76000,  84000,  92500,  101500, 110000, 120000, 130000, 140000, 150000,
+    160000, 170000, 180000, 190000, 200000, 210000, 222000, 234000, 246000,
+    258000, 270000, 282000, 294000, 306000, 318000, 330000, 344000, 358000,
+    372000, 386000, 400000, 414000, 428000, 442000, 456000, 470000, 486000,
+    502000, 518000, 534000, 550000, 566000, 582000, 598000, 614000, 630000,
+    648000, 666000, 684000, 702000, 720000, 738000, 756000, 774000, 792000,
+    810000, 830000, 850000, 870000, 890000, 910000, 930000, 950000, 970000,
+    999999,
+};
+
+// clang-format on
+
+__asm__(".text\n.word 0\n.data");
