+++ src/saturn/game/cfgrw1.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/game/data", d60846A8, d_060846A8);
