+++ src/saturn/game/cfgrw2.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/game/data", d60854D8, d_060854D8);
