+++ src/saturn/game/cfgstr.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/game/data", d606B6D4, d_0606B6D4);
