+++ src/saturn/game/hdrstub.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/game/data", d6066030, d_06066030);
