+++ src/saturn/game/jewel.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u8 g_JewelSwordAttackBonus[] = {1, 3, 6, 10, 16, 23, 32, 0};
