+++ src/saturn/game/joseph.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u16 g_JosephsCloakColors[4] = {0x8000, 0x8000, 0x8000, 0x8000};
