+++ src/saturn/game/miscend.c
@@ -1,0 +1,39 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+char* g_LuckCode = "X-X!V''Q";
+char* g_AxeArmorCode = "AXEARMOR";
+char* g_GTIClubCode = "GTI-CLUB";
+
+__asm__(".text\n.byte 0\n.word 0x0009\n.data");
+
+u16 D_06085C9C[] = {
+    0x0005, 0x0001, 0x0006, 0x0002, 0x001A, 0x0005, 0x001A, 0x0005, 0x001A,
+    0x0005, 0x0007, 0x0000, 0x0007, 0x0000, 0x0007, 0x0000, 0x0000, 0x0000,
+    0x0002, 0x0018, 0x0110, 0x0016, 0x0004, 0x0074, 0x011F, 0x0072, 0x0129,
+    0x0072, 0x0012, 0x001D, 0x001B, 0x001D, 0x0014, 0x0029, 0x0028, 0x0016,
+    0x1429, 0x142D, 0x1224, 0x1425, 0x1128, 0x1425, 0x1425, 0x192E, 0x132A,
+    0x1125, 0x1429, 0x0005, 0x0001, 0x0011, 0x0000, 0x0011, 0x000A, 0x0006,
+    0x0000, 0x0006, 0x0000, 0x0006, 0x0000, 0x0006, 0x0000, 0x0000, 0x0000,
+    0x0002, 0x0018, 0x0004, 0x0074, 0x0019, 0x0074, 0x0010, 0x001D, 0x0018,
+    0x001D, 0x0020, 0x001D, 0x0014, 0x0029, 0x002D, 0x0015, 0x0000, 0x1429,
+    0x142D, 0x1224, 0x1425, 0x1128, 0x1425, 0x1425, 0x192E, 0x132A, 0x1125,
+    0x1429, 0x0000, 0x0000, 0x0064, 0x0000, 0x0064, 0x0000, 0x0000, 0x0000,
+    0x0000, 0x0000, 0x0008, 0xC0A8, 0x8421, 0x84AB, 0x8D2F, 0x91D6, 0x929B,
+    0x994A, 0xA5CE, 0x8421, 0x9CE7, 0xFFE0, 0x98A5, 0xA94A, 0xBE2E, 0xC2F7,
+    0xF39C, 0x2030, 0x0000, 0x1408, 0x0300, 0x1408, 0x0328, 0x0C08, 0x0350,
+    0x0C08, 0x0368, 0x181C, 0x0380, 0x141C, 0x0428, 0x0C1C, 0x04B4, 0x1410,
+    0x0508, 0x1010, 0x0558, 0x0C10, 0x0598, 0x1410, 0x05C8, 0x1014, 0x0618,
+    0x1010, 0x0668, 0x1010, 0x06A8, 0x1010, 0x06E8, 0x0000, 0x0728, 0x0607,
+    0x840C, 0x0607, 0x8404, 0x0607, 0x83F8, 0x0607, 0x83F0, 0x0607, 0x83E8,
+    0x0607, 0x83E0, 0x0607, 0x83D4, 0x0607, 0x83CC, 0x0607, 0x83C8, 0x0607,
+    0x83C0, 0x0607, 0x83BC, 0x0607, 0x83B4, 0x0607, 0x83AC, 0x0607, 0x83A4,
+    0x0607, 0x839C, 0x0607, 0x8394, 0x0607, 0x8390, 0x0607, 0x838C, 0x0607,
+    0x8380, 0x0607, 0x8374, 0x0607, 0x8368, 0x0607, 0x8364, 0x0607, 0x8360,
+    0x0607, 0x8358, 0x0607, 0x8350, 0x0607, 0x8350, 0x0607, 0x8348, 0x0607,
+    0x833C, 0x0607, 0x849C, 0x0607, 0x848C, 0x0607, 0x846C, 0x0607, 0x844C,
+    0x0607, 0x8434, 0x0607, 0x8414, 0x0000, 0x0000, 0x000A, 0x000F, 0x000B,
+    0x0010, 0x0009, 0x000E, 0x000D, 0x0012, 0x0007, 0x000C, 0x0008, 0x000D,
+    0x000C, 0x0011, 0x0000, 0x0000, 0x0607, 0xAC40, 0x0000, 0x0000, 0x0000,
+    0x0000, 0x00FF,
+};
