+++ src/saturn/game/mnstate.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s32 D_801375C8 = 0;
