+++ src/saturn/game/savemsg.c
@@ -1,0 +1,26 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u8 g_BackupRamMessages[] =
+    "カートリッジＲＡＭ\0\0"
+    "本体ＲＡＭ\0\0"
+    "いいえ\0\0"
+    "はい\0\0\0\t"
+    "バックアップＲＡＭを\x01選んで下さい。\0"
+    "の全ての記録を\0\0"
+    "消します。（初期化します）\x01"
+    "　　　　よろしいですか？\0"
+    "　　　記録の更新をしますか？\0\0\0\t"
+    "　　　　セーブしますか？\0\0\0\t"
+    "セーブしました。\0\0\0\t"
+    "　データが５つ全てセーブされて　\x01"
+    "　います。\0"
+    "消すことができませんでした。\x01"
+    "（初期化できませんでした）\0"
+    "のバックアップ\0\0"
+    "の準備がまだできていません。\x01"
+    "（初期化されていません）\0\0\t"
+    "このゲームの記録をセーブするため\x01"
+    "には７０の空き容量が必要です。\0"
+    "記録を正しくセーブすることができ\x01"
+    "ませんでした。";
