+++ src/saturn/game_1.c
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+#include "sattypes.h"
+#include "game.h"
+
+void PlaySfx(s32 sfxId);
+
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6066B30, func_06066B30);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6066B74, func_06066B74);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6066CE0, func_06066CE0);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6066FE0, func_06066FE0);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6067090, func_06067090);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f60674B8, func_060674B8);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6067958, func_06067958);
+INCLUDE_ASM("asm/saturn/game/data", d606797C, d_0606797C);
+INCLUDE_ASM("asm/saturn/game/f_nonmat", f6067A04, func_06067A04);
