+++ src/saturn/ric/anims.c
@@ -1,0 +1,131 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+AnimationFrame ric_anim_stand[] = {
+    POSE(7, 1, 1),  POSE(7, 2, 1),  POSE(7, 3, 1),  POSE(7, 2, 1),
+    POSE(7, 1, 1),  POSE(7, 2, 1),  POSE(7, 3, 1),  POSE(7, 2, 1),
+    POSE(7, 1, 1),  POSE(7, 2, 1),  POSE(7, 3, 1),  POSE(7, 2, 1),
+    POSE(7, 1, 1),  POSE(7, 2, 1),  POSE(7, 3, 1),  POSE(7, 2, 1),
+    POSE(7, 1, 1),  POSE(7, 2, 1),  POSE(7, 3, 1),  POSE(7, 2, 1),
+    POSE(8, 1, 1),  POSE(8, 2, 1),  POSE(8, 3, 1),  POSE(8, 2, 1),
+    POSE(9, 1, 1),  POSE(9, 2, 1),  POSE(9, 3, 1),  POSE(9, 2, 1),
+    POSE(10, 1, 1), POSE(10, 2, 1), POSE(10, 3, 1), POSE(10, 2, 1),
+    POSE(11, 1, 1), POSE(11, 2, 1), POSE(11, 3, 1), POSE(11, 2, 1),
+    POSE(12, 1, 1), POSE(12, 2, 1), POSE(13, 3, 1), POSE(13, 2, 1),
+    POSE(15, 1, 1), POSE(15, 2, 1), POSE(15, 3, 1), POSE(128, 2, 1),
+    POSE(4, 4, 1),  POSE(4, 5, 1),  POSE(4, 6, 1),  POSE(64, 7, 1),
+    POSE_END};
+AnimationFrame ric_anim_stand_relax[] = {POSE(64, 7, 1), POSE_LOOP(0)};
+AnimationFrame ric_anim_walk[] = {
+    POSE(3, 8, 1),  POSE(6, 9, 1),  POSE(5, 10, 1), POSE(6, 11, 1),
+    POSE(7, 12, 1), POSE(6, 13, 1), POSE(5, 14, 1), POSE(6, 15, 1),
+    POSE(4, 8, 1),  POSE_LOOP(0)};
+AnimationFrame ric_anim_crouch_from_stand2[] = {
+    0x0001, 0x0410, 0x0001, 0x0611, 0x0010, 0x0612, POSE_END};
+AnimationFrame ric_anim_crouch[] = {POSE(64, 18, 3), POSE_END};
+AnimationFrame ric_anim_land_from_air_run[] = {POSE(16, 152, 3), POSE_END};
+AnimationFrame ric_anim_crouch_from_stand[] = {
+    0x0001, 0x0410, 0x0001, 0x0611, 0x0001, 0x0612, POSE_END};
+AnimationFrame D_801554E0[] = {
+    0x0001, 0x0612, 0x0001, 0x0611, 0x0001, 0x0410, POSE_END};
+AnimationFrame D_801554F0[] = {
+    0x0008, 0x0813, 0x0003, 0x0415, 0x0005, 0x0416,  0x0003,
+    0x0415, 0x0006, 0x0417, 0x0050, 0x0418, POSE_END};
+AnimationFrame D_8015550C[] = {
+    0x0008, 0x0814, 0x0003, 0x0415, 0x0005, 0x0416,  0x0003,
+    0x0415, 0x0006, 0x0417, 0x0050, 0x0418, POSE_END};
+AnimationFrame D_80155528[] = {0x0006, 0x0417, 0x0050, 0x0418, POSE_END};
+AnimationFrame D_80155534[] = {
+    0x0002, 0x0813, 0x0006, 0x0417, 0x0050, 0x0418, POSE_END};
+AnimationFrame D_80155544[] = {0x0004, 0x0820, 0x0003, 0x0C21, 0x0002,  0x0C22,
+                               0x0010, 0x0C22, 0x0001, 0x0C22, POSE_END};
+AnimationFrame D_8015555C[] = {
+    0x0004, 0x0820, 0x0003, 0x0C21, 0x0003, 0x0C22, 0x0003,
+    0x0C23, 0x0002, 0x0C24, 0x0001, 0x0A1C, 0x0001, 0x0A1B,
+    0x0020, 0x0425, 0x0012, 0x0819, 0x0030, 0x0418, POSE_END};
+AnimationFrame D_80155588[] = {
+    0x0002, 0x0226, 0x0002, 0x0227, 0x0002, 0x0228, 0x0002,  0x0229,
+    0x0002, 0x022A, 0x0002, 0x022B, 0x0010, 0x022C, POSE_END};
+AnimationFrame D_801555A8[] = {
+    0x0002, 0x062F, 0x0002, 0x0630, 0x0002, 0x0631, 0x0002,  0x0632,
+    0x0002, 0x0633, 0x0002, 0x0634, 0x0010, 0x0635, POSE_END};
+AnimationFrame D_801555C8[] = {
+    0x0002, 0x0836, 0x0002, 0x0837, 0x0002, 0x0428, 0x0002,  0x0429,
+    0x0002, 0x042A, 0x0002, 0x042B, 0x0010, 0x042C, POSE_END};
+AnimationFrame D_801555E8[] = {
+    0x0002, 0x0226, 0x0002, 0x0227, 0x0002, 0x0228, 0x0002,  0x0229,
+    0x0003, 0x022A, 0x0010, 0x022D, 0x0006, 0x022E, POSE_END};
+AnimationFrame ric_anim_throw_daggers[] = {
+    0x0002, 0x0226, 0x0002, 0x0227, 0x0002, 0x0228, 0x0002,      0x0229,
+    0x0002, 0x022A, 0x0006, 0x022D, 0x0002, 0x022E, 0x0002,      0x0228,
+    0x0002, 0x0229, 0x0003, 0x022A, 0x0002, 0x022D, POSE_LOOP(7)};
+AnimationFrame D_80155638[] = {
+    0x0002, 0x0836, 0x0002, 0x0837, 0x0002, 0x0428, 0x0002,  0x0429,
+    0x0003, 0x042A, 0x0010, 0x042D, 0x0006, 0x042E, POSE_END};
+AnimationFrame ric_anim_stand_in_air[] = {
+    POSE(5, 19, 0),  POSE(1, 26, 0),   POSE(1, 25, 0),
+    POSE(6, 125, 0), POSE(32, 126, 0), POSE_LOOP(4)};
+AnimationFrame ric_anim_run[] = {
+    0x0003, 0x1281, 0x0003, 0x1282, 0x0003, 0x1283, 0x0003, 0x1284,      0x0003,
+    0x1285, 0x0003, 0x1286, 0x0003, 0x127F, 0x0003, 0x1280, POSE_LOOP(0)};
+AnimationFrame ric_anim_stun[] = {POSE(64, 32, 4), POSE_END};
+AnimationFrame D_8015569C[] = {
+    0x0001, 0x02CE, 0x0001, 0x02CF, 0x0001,  0x02CE, 0x0003,
+    0x02CF, 0x0003, 0x02CE, 0x0005, 0x02CF,  0x0001, 0x0226,
+    0x0001, 0x022E, 0x0020, 0x0202, POSE_END};
+AnimationFrame D_801556C4[] = {
+    0x0001, 0x08B7, 0x0002, 0x08B8, 0x0001, 0x08B6, 0x0001,  0x08B7,
+    0x0001, 0x08B9, 0x0001, 0x08B6, 0x0001, 0x08B7, 0x0001,  0x08BA,
+    0x0001, 0x08B6, 0x0001, 0x08B7, 0x0001, 0x08BC, 0x0001,  0x08B7,
+    0x0001, 0x08BB, 0x0001, 0x08B7, 0x0020, 0x08B6, POSE_END};
+AnimationFrame D_80155704[] = {
+    0x0001, 0x06C8, 0x0001, 0x06C9, 0x0001, 0x06C8, 0x0002,
+    0x06C9, 0x0002, 0x06C8, 0x0003, 0x06C9, 0x0003, 0x06C8,
+    0x0005, 0x06C9, 0x0001, 0x062F, 0x0001, 0x0611, POSE_END};
+AnimationFrame ric_anim_brandish_whip[] = {0x0080, 0x02D0, POSE_END};
+AnimationFrame D_80155738[] = {0x0080, 0x06D1, POSE_END};
+AnimationFrame D_80155740[] = {0x0080, 0x08D2, POSE_END};
+AnimationFrame D_80155748[] = {0x0080, 0x0011, POSE_END};
+AnimationFrame D_80155750[] = {
+    0x0002, 0x0218, 0x0002, 0x0487, 0x0002, 0x0488, 0x0001,
+    0x0642, 0x0008, 0x1489, 0x0010, 0x148A, 0x0008, 0x1489,
+    0x0002, 0x0642, 0x0002, 0x0641, 0x0002, 0x0611, POSE_END};
+AnimationFrame D_8015577C[] = {0x0016, 0x228C, 0x0080, 0x228D, POSE_END};
+AnimationFrame D_80155788[] = {
+    0x0003, 0x081B, 0x0003, 0x081A, 0x0003, 0x0819, 0x0030, 0x0418, POSE_END};
+AnimationFrame ric_anim_high_jump[] = {
+    0x0001, 0x089B, 0x0001, 0x089C, 0x0010, 0x089B, POSE_END};
+AnimationFrame ric_anim_point_finger_up[] = {
+    0x0002, 0x00A4, 0x0002, 0x00A5, 0x0001, 0x00A4, 0x0001,
+    0x00A5, 0x000D, 0x00A4, 0x0002, 0x00A6, 0x0003, 0x00A7,
+    0x0002, 0x00A6, 0x0080, 0x00A6, 0x0002, 0x0026, POSE_END};
+AnimationFrame ric_anim_blade_dash[] = {
+    0x0001, 0x0008, 0x0001, 0x0038, 0x0001, 0x0092,  0x0001, 0x0093, 0x0001,
+    0x0094, 0x0001, 0x0095, 0x0001, 0x0096, 0x0001,  0x0097, 0x0001, 0x0094,
+    0x0001, 0x0095, 0x0001, 0x0096, 0x0001, 0x0097,  0x0001, 0x0093, 0x0001,
+    0x0094, 0x0001, 0x0095, 0x0001, 0x0096, 0x0001,  0x0097, 0x0010, 0x0092,
+    0x0002, 0x0037, 0x0001, 0x0090, 0x0002, 0x0091,  0x0003, 0x0090, 0x0004,
+    0x0091, 0x000C, 0x0090, 0x0004, 0x0010, POSE_END};
+AnimationFrame ric_anim_flame_whip[] = {
+    0x0004, 0x02A2, 0x0001,  0x02A8, 0x0002, 0x02A9, 0x0018, 0x02A8, 0x0004,
+    0x02AC, 0x0001, 0x02AA,  0x0002, 0x02AB, 0x0018, 0x02AA, 0x0008, 0x02AC,
+    0x0005, 0x02AD, 0x0004,  0x02AC, 0x0003, 0x02AD, 0x0002, 0x02AE, 0x0002,
+    0x02AD, 0x0001, 0x02AE,  0x0001, 0x02AF, 0x0001, 0x02AE, 0x000C, 0x02AF,
+    0x0001, 0x02B0, 0x0002,  0x02B1, 0x0001, 0x02B2, 0x0001, 0x02B3, 0x0001,
+    0x02B4, 0x0001, 0x02B5,  0x0001, 0x02B4, 0x0001, 0x02B3, 0x0001, 0x02B2,
+    0x0030, 0x02B0, POSE_END};
+AnimationFrame D_801558B4[] = {
+    0x0005, 0x0013, 0x0001, 0x001A, 0x0001, 0x0019, 0x0006,  0x007D,
+    0x0083, 0x007E, 0x0080, 0x007E, 0x0080, 0x0019, POSE_END};
+AnimationFrame D_801558D4[] = {0x000C, 0x0019, POSE_END};
+AnimationFrame D_801558DC[] = {0x0006, 0x001A, POSE_END};
+AnimationFrame D_801558E4[] = {
+    0x0002, 0x0237, 0x0001, 0x0290, 0x0002, 0x0291, 0x0003, 0x0290, 0x0004,
+    0x0291, 0x0005, 0x0291, 0x0005, 0x0290, 0x0005, 0x0291, 0x0005, 0x0291,
+    0x0005, 0x0290, 0x0005, 0x0291, 0x000C, 0x0290, 0x0010, 0x0226, POSE_END};
+AnimationFrame D_8015591C[] = {
+    0x0002, 0x00A4, 0x0002, 0x00A5, 0x0001, 0x00A4, 0x0001,  0x00A5, 0x000D,
+    0x00A4, 0x0002, 0x00A6, 0x0003, 0x00A7, 0x0002, 0x00A6,  0x0080, 0x00A6,
+    0x0080, 0x00A6, 0x0080, 0x00A6, 0x0080, 0x00A6, POSE_END};
+AnimationFrame D_80155950[] = {
+    0x0080, 0x00A6, 0x0080, 0x00A6, 0x0080, 0x00A6, 0x0002, 0x0026, POSE_END};
