+++ src/saturn/ric/c2f40.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+u16 D_060C2F40 = 0xFFFF;
