+++ src/saturn/ric/globdat.c
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s32 DAT_060c4118 = 0;
+s32 DAT_060c411c = 0;
