+++ src/saturn/ric/gprolog.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s32 g_IsPrologueStage = 0;
