+++ src/saturn/ric/hdrstub.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/richter/data", d60A5044, d_060A5044);
