+++ src/saturn/ric/ptrtbl.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s16 PTR_DAT_060cd7f0[] = {8, 8, 14, 13, 20, 18, 23, 144};
