+++ src/saturn/ric/rictl10.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/richter/data", d60CD800, d_060CD800);
