+++ src/saturn/ric/rictl7.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/richter/data", d60C3982, d_060C3982);
