+++ src/saturn/ric/rictl9.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "inc_asm.h"
+
+INCLUDE_ASM("asm/saturn/richter/data", d60CD750, d_060CD750);
