+++ src/saturn/ric/savevar.c
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s32 d_060cd748 = 0;
+s32 d_060cd74c = 0;
