+++ src/saturn/ric/whipdat.c
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "sattypes.h"
+
+s16 D_80155D30[] = {
+    0x10, 0x18, 0x11, 0x19, 0x12, 0x1A, 0x13, 0x1B,
+    0x14, 0x1C, 0x15, 0x1D, 0x16, 0x1E, 0x17,
+};
+
+u16 unused_2[] = {0x02, 0x01, 0x02, 0x02, 0x02, 0x03, 0x02, 0x04, 0x00};
